### 2019-02-02 4a196e4
fixing bibtex errors
