+++ paper.bib
@@ -75,7 +75,7 @@
 	          {Whitney}, A.~R. and {Williams}, A. and {Williams}, C.~L. and
 	          {Wyithe}, J.~S.~B.},
   title = "{Fast Holographic Deconvolution: A New Technique for Precision Radio Interferometry}",
-  journal = {\apj},
+  journal = {ApJ},
   archivePrefix = "arXiv",
   eprint = {1209.1653},
   primaryClass = "astro-ph.IM",
@@ -87,7 +87,7 @@
   pages = {17},
   doi = {10.1088/0004-637X/759/1/17},
   adsurl = {http://adsabs.harvard.edu/abs/2012ApJ...759...17S},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
   url = {https://github.com/EoRImaging/FHD}
 }
 
@@ -112,10 +112,10 @@ Bdsk-Url-1 = {https://dx.doi.org/10.1051/0004-6361/201322068}}
 @article{astropy:2018,
 Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 Adsurl = {https://ui.adsabs.harvard.edu/#abs/2018AJ....156..123T},
-Author = {{Price-Whelan}, A.~M. and {Sip{'{o}}cz}, B.~M. and {G{"u}nther}, H.~M. and {Lim}, P.~L. and {Crawford}, S.~M. and {Conseil}, S. and {Shupe}, D.~L. and {Craig}, M.~W. and {Dencheva}, N. and {Ginsburg}, A. and {VanderPlas}, J.~T. and {Bradley}, L.~D. and {P{'e}rez-Su{'a}rez}, D. and {de Val-Borro}, M. and {Paper Contributors}, (Primary and {Aldcroft}, T.~L. and {Cruz}, K.~L. and {Robitaille}, T.~P. and {Tollerud}, E.~J. and {Coordination Committee}, (Astropy and {Ardelean}, C. and {Babej}, T. and {Bach}, Y.~P. and {Bachetti}, M. and {Bakanov}, A.~V. and {Bamford}, S.~P. and {Barentsen}, G. and {Barmby}, P. and {Baumbach}, A. and {Berry}, K.~L. and {Biscani}, F. and {Boquien}, M. and {Bostroem}, K.~A. and {Bouma}, L.~G. and {Brammer}, G.~B. and {Bray}, E.~M. and {Breytenbach}, H. and {Buddelmeijer}, H. and {Burke}, D.~J. and {Calderone}, G. and {Cano Rodr{'i}guez}, J.~L. and {Cara}, M. and {Cardoso}, J.~V.~M. and {Cheedella}, S. and {Copin}, Y. and {Corrales}, L. and {Crichton}, D. and {D{ extquoteright}Avella}, D. and {Deil}, C. and {Depagne}, {'E}. and {Dietrich}, J.~P. and {Donath}, A. and {Droettboom}, M. and {Earl}, N. and {Erben}, T. and {Fabbro}, S. and {Ferreira}, L.~A. and {Finethy}, T. and {Fox}, R.~T. and {Garrison}, L.~H. and {Gibbons}, S.~L.~J. and {Goldstein}, D.~A. and {Gommers}, R. and {Greco}, J.~P. and {Greenfield}, P. and {Groener}, A.~M. and {Grollier}, F. and {Hagen}, A. and {Hirst}, P. and {Homeier}, D. and {Horton}, A.~J. and {Hosseinzadeh}, G. and {Hu}, L. and {Hunkeler}, J.~
[…]
