### 2019-02-04 #1234 @whedon
[REVIEW]: pyuvsim: a comprehensive simulation package for radio interferometers in python.
**Submitting author:** @aelanman (<a href="http://orcid.org/0000-0003-2116-3573">Adam Lanman</a>)
**Repository:** <a href="https://github.com/RadioAstronomySoftwareGroup/pyuvsim/" target ="_blank">https://github.com/RadioAstronomySoftwareGroup/pyuvsim/</a>
**Version:** v0.2.1
