### 2019-02-06 #1234 @ygrange
Just going through the checkboxes I found two comments:
- Please add doi to the Dalcin, Paz and Storti paper ( https://doi.org/10.1016/j.jpdc.2005.03.010 )
- There seems to be no tag for version 0.2.1 (there is a 0.2.0). I don't referee often enough for JOSS to know what the exact policy here is: to have the version released after my comments or to have a release now and add a new one at publication time if the code needs to change because of sometning I said. Maybe the editor should comment on this.
