### 2019-02-06 #1234 @arfon
I think it's OK to review what is in the repository now (i.e. `master`) and then for there to be a new release before publication.

Perhaps @aelanman could clarify what they're asking to be reviewed.
