### 2019-02-06 #1234 @aelanman
Hello, @ygrange . We're asking for the contents of `master` to be reviewed. We've issued releases for every major version change (generation.major.minor) so far, but I assume we'll do another release after the review is finished.

I'll add the missing doi.

Thanks!
