### 2019-02-06 80fd8d2
fix missing doi and journal abbreviation
