+++ paper.bib
@@ -99,7 +99,7 @@ Author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. 
 Doi = {10.1051/0004-6361/201322068},
 Eid = {A33},
 Eprint = {1307.6212},
-Journal = {aap},
+Journal = {AAP},
 Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
 Month = oct,
 Pages = {A33},
@@ -130,6 +130,7 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}
 title={{MPI} for {Python}},
 author={Dalc{\'\i}n, Lisandro and Paz, Rodrigo and Storti, Mario},
 journal={Journal of Parallel and Distributed Computing},
+doi={10.1016/j.jpdc.2005.03.010}
 volume={65},
 number={9},
 pages={1108--1115},
