+++ paper.bib
@@ -130,7 +130,7 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}
 title={{MPI} for {Python}},
 author={Dalc{\'\i}n, Lisandro and Paz, Rodrigo and Storti, Mario},
 journal={Journal of Parallel and Distributed Computing},
-doi={10.1016/j.jpdc.2005.03.010}
+doi={10.1016/j.jpdc.2005.03.010},
 volume={65},
 number={9},
 pages={1108--1115},
