### 2019-02-06 6ae7e6c
Check beam kernel width in 2D instead of 1. Getting a tophat now
