### 2019-02-08 #1234 @aelanman
I just noticed that some authors' first names were cut from citations. I'll have to fix this, but I'm going to wait for more feedback first.
