### 2019-02-08 ec8c723
fix changelog -- was missing v0.2.1
