### 2019-02-09 7d1ac44
Fri Feb  8 20:13:59 EST 2019 scbt from Zacs memo
