### 2019-02-09 9034896
various fixes for pypi packaging
