+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 - Support for only simulating one baseline per redundant group
+- Checking the beam kernel width makes sense for Airy beams
 
 ### Fixed
 - Using the correct Bessel function in the definition of the Airy beam
@@ -23,7 +24,6 @@
 - Coarse horizon cut for reduced computational load.
 - Selection keys in obsparam files (valid options for UVData.select)
 - New batch scripts for profiling jobs on a SLURM-managed cluster, and plotting the results.
-- Checking the beam kernel width makes sense for Airy beams
 
 ### Changed
 - Rename write_uvfits to write_uvdata
