### 2019-02-10 d8f8eca
Frequency array values specify channel centers
