+++ pyuvsim/data/test_config/param_10time_10chan_1.yaml
@@ -2,9 +2,10 @@ filing:
   outdir: './tempdir/'
   outfile_prefix: 'sim'
 freq:
-  freq_array : [1.00000000e+08,   1.00080000e+08,   1.00160000e+08, 1.00240000e+08, 
-              1.00320000e+08,   1.00400000e+08,   1.00480000e+08, 1.00560000e+08,
-              1.00640000e+08, 1.00720000e+08]
+  freq_array: [100.5, 101.5, 102.5, 103.5, 104.5, 105.5, 106.5, 107.5, 108.5, 109.5]
+  #  freq_array : [1.00000000e+08,   1.00080000e+08,   1.00160000e+08, 1.00240000e+08, 
+  #            1.00320000e+08,   1.00400000e+08,   1.00480000e+08, 1.00560000e+08,
+  #            1.00640000e+08, 1.00720000e+08]
 sources:
   catalog: 'mock'
   mock_arrangement: 'zenith'
