### 2019-02-10 93aa562
Fix time array
