+++ pyuvsim/data/test_config/param_10time_10chan_3.yaml
@@ -13,6 +13,6 @@ telescope:
   telescope_config_name: '28m_triangle_10time_10chan.yaml'
 time:
   Ntimes: 10
-  end_time: 2457458.175040789
+  end_time: 2457458.1750407889
   integration_time: 11.0
   start_time: 2457458.1738949567
