+++ pyuvsim/data/test_config/param_10time_10chan_4.yaml
@@ -12,6 +12,6 @@ telescope:
   array_layout: 'triangle_bl_layout.csv'
   telescope_config_name: '28m_triangle_10time_10chan.yaml'
 time:
-  end_time: 2457458.175040789
+  end_time: 2457458.1750407889
   integration_time: 11.0
   start_time: 2457458.1738949567
