### 2019-02-10 90098fd
Update documentation to clarify frequency/time conventions
