### 2019-02-13 e02f548
Change uvdata_to_config for new channel definition
