### 2019-02-15 a0a5995
Docs on time array
