### 2019-02-15 #1234 @ygrange
Just letting you know I got a bit delayed because I had some issues installing mpi4py on my centos test container. Now I have a running version and I can have a look at functionality.
