### 2019-02-15 8139d9c
Redefining kernel_width test
