### 2019-02-23 6774083
Fix test to still use sigma and correctly make analytic beam
