+++ pyuvsim/data/test_config/param_10time_10chan_2.yaml
@@ -10,7 +10,7 @@ sources:
   mock_arrangement: 'zenith'
 telescope:
   array_layout: 'triangle_bl_layout.csv'
-  telescope_config_name: '28m_triangle_10time_10chan.yaml'
+  telescope_config_name: '28m_triangle_10time_10chan_sigmaonly.yaml'
 time:
   Ntimes: 10
   duration_days: 0.001145833
