+++ pyuvsim/data/test_config/28m_triangle_10time_10chan_sigmaonly.yaml
@@ -1,0 +1,8 @@
+beam_paths:
+  0: HERA_NicCST.uvbeam
+  1: 'uniform'
+  2: 'gaussian'
+  3: 'gaussian'
+sigma: 0.02
+telescope_location: (-30.721527777777847, 21.428305555555557, 1073.0000000046566)
+telescope_name: HERA
