### 2019-02-23 e881067
Fix test to still use sigma and correctly make analytic beam
