### 2019-02-23 ec2f56f
Fix sigma/diameter scaling relation
