### 2019-02-23 0689a40
Test that Airy beam and chromatic gaussian are close within first null
