+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Added
 - Checking the beam kernel width makes sense for Airy beams
+- Enabled Gaussian beams to be defined from antenna diameter and have chromaticity
 
+### Changed
+- Pending deprecation of gaussian beams defined from sigma parameter
 
 ## [0.2.3] - 2019-2-17
 
