### 2019-02-23 57497bb
Update docs to describe beam parameters
