### 2019-02-23 42513d9
Check error case for gaussian beam
