### 2019-02-23 a96d7c7
Remove unused check
