### 2019-03-05 34a702f
Remove middle initials from pyuvdata citation. This should fix the inline citations
