### 2019-03-08 #1234 @ygrange
I had some things taking up priority recently. Applogies! I have some
draft comments but I want to run through examples, read the docs to see
if I miss something, etc. I’ll plan time for it early next week
(mo/tue).

--
Y. Grange

> Op 8 mrt. 2019 om 17:41 heeft Adam Lanman <notifications@github.com>
het volgende geschreven:
>
> @arfon
> Hi Arfon. It's been quite some time... can my collaborators and I
expect a review soon?
>
