### 2019-03-12 #1234 @ygrange
Here it comes (again, with apologies for the delay)...

In general: I like the tool and the ambition it has to be a tool to work through the measurement equation without short cuts. I have several comments of which most are textual.

**Paper**
* 1 minor comment: you list future low-frequency telescopes. Any reason you don't list LOFAR which is a current low-frequency radio telescope? I think such a tool could be very useful for current telescopes as well (for algorithm development and other R&D on the array).

* In the readthedocs, you give a nice list of other tools that simulate radio data. You then mention what sets you apart from them (i.e. they all use apporximations to speed up). I would add this to the paper and if possible, would it be possible to
quantify the effect of your fully analytic approach on the performance and precision (do I gain 10% precision in 200% of the run time; these kind of numbers)? This is the only comment that would require a bit of extra work. I am not really sure what the policy of JOSS is in this respect and whether this is in the category "nice to have" or "essential" in that case (@arfon: any comments on this?).
* Also it would be nice if either the paper or the documentation could give potentially interested users a rough idea whether using this tool is the best fit to their use case as opposed to CASA, OSCAR, FHD, PRISim, etc.

* "The measurement equation": add citation (Hamaker et al. 1996 ).

**README.md**
* The README has a section "Inputs" but doesn't seem to mention outputs (It's a standard uvfits I guess so it would be a single line)

* In both the README and the documentation the following quick-run command is given:
`mpirun -n 20 python run_param_pyuvsim.py reference_simulations/obsparam_1.1.yaml`
It should be
`mpirun -n 20 python scripts/run_param_pyuvsim.py reference_simulations/obsparam_ref_1.1.yaml`
and specify where to execute it from
* Community guidelines are present, but it doesn't really say where one can get support (questions that are not per se bug reports). Is that via github issues?

**readthedocs documentation**
* "Parameter and configuration Files" section doesn't state where all the files should be. I assume the paths are relative to the path of the main yaml file but It's best to explicitly write that down.

* In the "Enabling Profiling" section
`To run a simulation with profiling enabled, run the command profiling.set_profiler before starting`
change to`(...) run the command profiling.set_profiler() before `

**installation instruction and procedure**
* Installation: you only present how to build the master branch. Since you have releases, I think you should explicitly document hoe to build a (the latest?) release.
* I would strongly advice to copy the installation instructions from readthedocs to the README.md file too.
* Adding a requirements.txt could also make sense although I did notice that pyuvsim doesn't really do requirement installation well.
* Also listing max an min versions if applicable (seems mpi4py 1.3.1 is too old for example; this is the version from Centos7 EPEL).

**tests**
* Is there a reason you don't add the availability of nosetests to the documentation? If I know a package is testable, I will test it especially if it documents how to test it.
* One test fails if h5py is not installed. Probably it is a dependency via pyuvdata, but it's worth documenting that you need h5py to pass all tests.
* Also if you run the tests twice, the `test_file.uvfits` file will already exist and the system will cope by creating `test_file_0.uvfits` making the corresponding test to fail.
