### 2019-03-17 62b158f
Add unit testing and output sections to README
