### 2019-03-17 a535c6d
fix usage.rst
