+++ paper.bib
@@ -138,3 +138,21 @@ year={2005},
 publisher={Elsevier}
 }
 
+
+@article{hamaker_understanding_1996,
+	title = {Understanding radio polarimetry. {I}. {Mathematical} foundations},
+	volume = {117},
+	copyright = {© European Southern Observatory (ESO), 1996},
+	issn = {0365-0138, 1286-4846},
+	url = {https://doi.org/10.1051/aas:1996146},
+	doi = {10.1051/aas:1996146},
+	abstract = {Astronomy and Astrophysics, Supplement Series (A\&AS) published data papers, either observational or theoretical, as well as extensive data material forming the basis of papers with astrophysical results},
+	language = {en},
+	number = {1},
+	urldate = {2017-10-09},
+	journal = {Astronomy and Astrophysics Supplement Series},
+	author = {Hamaker, J. P. and Bregman, J. D. and Sault, R. J.},
+	month = may,
+	year = {1996},
+	pages = {137--147}
+}
