### 2019-03-17 54183ee
Better VOTable handling via astropy Tables
