### 2019-03-18 ec2ceb7
attempted fix of ascii encoding bug
