+++ docs/make_index.py
@@ -6,6 +6,7 @@ Format the readme.md file into the sphinx index.rst file.
 """
 from __future__ import absolute_import, division, print_function
 
+import codecs
 import os
 import inspect
 import re
@@ -47,7 +48,6 @@ def write_index_rst(readme_file=None, write_file=None):
     out += ('\n\nFurther Documentation\n====================================\n'
             '.. toctree::\n'
             '   :maxdepth: 2\n\n'
-            '   :caption: Contents:\n'
             '   usage\n'
             '   parameter_files\n'
             '   classes\n')
@@ -57,6 +57,6 @@ def write_index_rst(readme_file=None, write_file=None):
     if write_file is None:
         write_path = os.path.dirname(os.path.abspath(inspect.stack()[0][1]))
         write_file = os.path.join(write_path, 'index.rst')
-    F = open(write_file, 'w')
+    F = codecs.open(write_file, 'w', 'utf-8')
     F.write(out)
     print("wrote " + write_file)
