### 2019-03-19 9542414
starting simulator comparison file
