+++ docs/make_index.py
@@ -48,6 +48,8 @@ def write_index_rst(readme_file=None, write_file=None):
     out += ('\n\nFurther Documentation\n====================================\n'
             '.. toctree::\n'
             '   :maxdepth: 2\n\n'
+            '   index\n'
+            '   comparison\n'
             '   usage\n'
             '   parameter_files\n'
             '   classes\n')
