### 2019-03-20 8d4b7f3
Update comparison.rst
Update PRISim features
