### 2019-03-20 5053169
Updated CASA parallel information
