### 2019-03-22 #1234 @aelanman
@ygrange Thank you for the feedback! I want to let you know that we've been working to address your comments. Most fixes have been made so far, but we're still trying to make a good comparison with other simulation tools.

It's difficult to quantify the performance in a meaningful way. Some of these simulators are very efficient along a particular axis but not others, and I'm not sure what to use as a consistent metric. We've started writing up a document comparing the architecture and features of pyuvsim with other simulators, as a way to recommend when it might be useful. [you can find the document here](https://github.com/RadioAstronomySoftwareGroup/pyuvsim/blob/joss_revisions/docs/comparison.rst)

If you'd like to take a look at the document, you can let us know if we're on the right track.

The other difficulty is in quantifying precision/accuracy. For our work, "accuracy" mostly means avoiding unwanted numerical artifacts, and it's hard to put a percentage on that. What has usually happened in the past is we develop a new method or drop an approximation, and by doing so uncover an unexpected effect. If you have any suggestions of a good metric to report, we can try to make a comparison. It would be easier to point to a few examples of the unwanted effects we've managed to avoid.

Perhaps it would be better to emphasize that pyuvsim strives to be as _universally applicable_ as possible. Since we don't make any assumptions about what numerical or instrumental effects are important to the user, we want the base functionality to support an exact calculation for any interferometer design. This is only possible by enabling a full sky, full-polarization calculation, with per-antenna user-defined beam models.
