### 2019-03-23 74f2712
Begin working on vectorized Sources objects
