+++ pyuvsim/simsetup.py
@@ -21,7 +21,7 @@ from astropy.coordinates import Angle, SkyCoord, EarthLocation, AltAz
 from pyuvdata import UVBeam, UVData
 import pyuvdata.utils as uvutils
 
-from .source import Source
+from .source import Sources
 from .analyticbeam import AnalyticBeam
 from .mpi import get_rank
 from .utils import check_file_exists_and_increment
@@ -80,15 +80,14 @@ def _config_str_to_dict(config_str):
     return param_dict
 
 
-def array_to_sourcelist(catalog_table, lst_array=None, time_array=None, latitude_deg=None, horizon_buffer=0.04364, min_flux=None, max_flux=None):
+def array_to_sourcelist(catalog_table, lst_array=None, latitude_deg=None, horizon_buffer=0.04364, min_flux=None, max_flux=None):
     """
-    Generate list of source.Source objects from a recarray, performing flux and horizon selections.
+    Generate a Sources object from a recarray, performing flux and horizon selections.
 
     Args:
         catalog_table: recarray of source catalog information. Must have columns with names
                        'source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency'
         lst_array: For coarse RA horizon cuts, lsts used in the simulation [radians]
-        time_array: Array of (float) julian dates corresponding with lst_array
         latitude_deg: Latitude of telescope in degrees. Used for declination coarse horizon cut.
         horizon_buffer: Angle (float, in radians) of buffer for coarse horizon cut. Default is about 10 minutes of sky rotation.
                         Sources whose calculated altitude is less than -horizon_buffer are excluded by the catalog read.
@@ -114,45 +113,48 @@ def array_to_sourcelist(catalog_table, lst_array=None, time_array=None, latitude
     if coarse_horizon_cut:
         lat_rad = np.radians(latitude_deg)
         buff = horizon_buffer
-    for (source_id, ra_j2000, dec_j2000, flux_I, freq) in catalog_table:
-        if min_flux:
-            if (flux_I < min_flux):
-                continue
-        if max_flux:
-            if (flux_I > max_flux):
-                continue
-        ra = Angle(ra_j2000, units.deg)
-        dec = Angle(dec_j2000, units.deg)
-        rise_lst = None
-        set_lst = None
-        if coarse_horizon_cut:
-            # Identify circumpolar sources and unrising sources.
-            tans = np.tan(lat_rad) * np.tan(dec.rad)
-            if tans < -1:
-                continue   # Source doesn't rise.
-            circumpolar = tans > 1
 
-            if not circumpolar:
-                rise_lst = ra.rad - np.arccos((-1) * tans)
-                set_lst = ra.rad + np.arccos((-1) * tans)
-                rise_lst -= buff
-                set_lst += buff
+    ra = Angle(catalog_table['ra_j2000'], units.deg)
+    dec = Angle(catalog_table['dec_j2000'], units.deg)
 
-                if rise_lst < 0:
-                    rise_lst += 2 * np.pi
-                if set_lst > 2 * np.pi:
-                    set_lst -= 2 * np.pi
+    if min_flux:
+        catalog_table = catalog_table[catalog_table['flux_I'] > min_flux]
 
-        source = Source(source_id, ra,
-                        dec,
-                        freq=freq * units.Hz,
-                        stokes=np.array([flux_I, 0., 0., 0.]),
-                        rise_lst=rise_lst,
-                        set_lst=set_lst)
+    if max_flux:
[…]
