+++ pyuvsim/source.py
@@ -9,15 +9,55 @@ from astropy.units import Quantity
 from astropy.time import Time
 from astropy.coordinates import Angle, SkyCoord, EarthLocation, AltAz
 
+from . import utils as simutils
+
 
 class Source(object):
     """
-    Defines a single point source at a given ICRS ra/dec coordinate, with a
-    flux density defined by stokes parameters.
+    Does nothing. Placeholder while I fix other things
+
+    """
+
+    def __init__(self):
+        self.nothing = None
+
+
+class _single_source(object):
+    """
+    Interface for accessing attributes of a single source component in a source list.
+    """
+
+    def __init__(self, sourcelist, index):
+        self.index = index
+        self.sourcelist = sourcelist
+
+    def __getattr__(self, key):
+
+        Ncomp_attrs = ['ra', 'dec', 'coherency_radec', 'coherency_local', 'stokes', 'alt_az', 'rise_lst', 'set_lst', 'freq', 'name']
+        scalar_attrs = ['time', 'pos_tol']
+
+        if key in Ncomp_attrs:
+            val = getattr(self.sourcelist, key)[..., self.index]
+            if isinstance(val, np.ndarray):
+                # Turn length 0 arrays into scalars
+                if val.size == 1:
+                    val = val.flatten()[0]
+            return val
+        else:
+            return getattr(self.sourcelist, key)
+
+
+class Sources(object):
+    """
+    Defines a set of point source components at given ICRS ra/dec coordinates, with a
+    flux densities defined by stokes parameters.
 
     Used to calculate local coherency matrix for source brightness in AltAz
     frame at a specified time.
     """
+
+    Ncomponents = None      # Number of point source components represented here.
+
     name = None
     freq = None
     stokes = None
@@ -25,6 +65,8 @@ class Source(object):
     dec = None
     coherency_radec = None
     alt_az = None
+    rise_lst = None
+    set_lst = None
 
     def __init__(self, name, ra, dec, freq, stokes, rise_lst=None, set_lst=None, pos_tol=np.finfo(float).eps):
         """
@@ -32,18 +74,19 @@ class Source(object):
 
         Args:
             name: Unique identifier for the source.
-            ra: astropy Angle object
+            ra: astropy Angle object, shape (Ncomponents,)
                 source RA in J2000 (or ICRS) coordinates
-            dec: astropy Angle object
+            dec: astropy Angle object, shape (Ncomponents,)
                 source Dec in J2000 (or ICRS) coordinates
             stokes:
+                shape (Ncomponents,4)
                 4 element vector giving the source [I, Q, U, V]
-            freq: astropy quantity
-                frequency of source catalog value
-            rise_lst: (float)
+            freq: astropy quantity, shape (Ncomponents,)
+                reference frequencies of flux values
+            rise_lst: (float), shape (Ncomponents,)
                 Approximate lst (radians) when the source rises. Set by coarse horizon cut in simsetup.
                 Default is None, meaning the source never rises.
-            set_lst: (float)
+            set_lst: (float), shape (Ncomponents,)
                 Approximate lst (radians) when the source sets.
                 Default is None, meaning the source never sets.
             pos_tol: float, defaults to minimum float in numpy
@@ -71,14 +114,45 @@ class Source(object):
         self.rise_lst = rise_lst
         self.set_lst = set_lst
 
+        self.Ncomponents = ra.size
+
+        assert np.all([self.Ncomponents == l for l in
+            [self.ra.size, self.dec.size, self.freq.size, self.stokes.shape[1]]]), 'Inconsistent quantity dimensions.'
+
+        if self.Ncomponents == 1:
[…]
