### 2019-03-24 54de069
circumpolar sources marked with nan instead of None
