### 2019-03-24 109cf53
update docstrings and formatting
