### 2019-03-26 #1234 @ygrange
I really like that document! That would make it much easier for me as a reader to appreciate the different tools around and why this is not "yet another tool".
I get the argument on precision/performance measurement. I think stating universal applicability is a good thing. It may also be a good idea to give a few examples of the effects you managed to avoid, especially if those actually occur in one or more of the mentioned alternative tools.

So I guess my main answer to your question is: yes, you are complerely on the right track here!
