### 2019-03-27 a071e8b
attempt to enable sources object subselection
