### 2019-03-27 072d747
make update_positions function
