### 2019-03-28 9c4de80
Fix docstring on parse_time_params
