### 2019-03-29 a4bc741
Check that end_freq - start_freq is divisible by channel_width
