### 2019-03-29 2b69e88
Avoid overwriting parameter dictionaries in time/freq parsers
