### 2019-03-29 c74b94d
Add utility functions to go from freq/time_array to parameters that generate it
