### 2019-03-29 c9599f0
Switch to safer ast.literal_eval function in simsetup
