### 2019-03-29 ce53403
Default to linear freq_interp for beams
