### 2019-03-29 6f99dce
Replace individual read_cst_beam calls in tests with one function
