+++ pyuvsim/tests/__init__.py
@@ -29,13 +29,16 @@ def compare_dictionaries(dic1, dic2):
     return bool(compare)
 
 
-def make_cst_beams():
+def make_cst_beams(freqs=None):
     beam = UVBeam()
     beam.freq_interp_kind = 'linear'
 
+    if freqs is None:
+        freqs = [150e6, 123e6]
+
     cst_files = ['HERA_NicCST_150MHz.txt', 'HERA_NicCST_123MHz.txt']
     beam_files = [os.path.join(DATA_PATH, 'NicCSTbeams', f) for f in cst_files]
-    beam.read_cst_beam(beam_files, beam_type='efield', frequency=[100e6, 123e6],
+    beam.read_cst_beam(beam_files, beam_type='efield', frequency=freqs,
                        telescope_name='HERA',
                        feed_name='PAPER', feed_version='0.1', feed_pol=['x'],
                        model_name='E-field pattern - Rigging height 4.9m',
