### 2019-03-30 dc62f2c
More informative responses for array assertion failure
