### 2019-03-30 ae33e01
Use time/freq_array_to_params functions for uvdata_to_config
