+++ pyuvsim/simsetup.py
@@ -699,26 +699,22 @@ def parse_time_params(time_params):
 
     _time_params = copy.deepcopy(time_params)
 
-    time_keywords = ['start_time', 'end_time', 'Ntimes', 'integration_time',
+    time_keywords = ['time_array', 'start_time', 'end_time', 'Ntimes', 'integration_time',
                      'duration_hours', 'duration_days']
-    st, et, nt, it, dh, dd = [tk in time_params for tk in time_keywords]
+    ta, st, et, nt, it, dh, dd = [tk in time_params for tk in time_keywords]
     kws_used = ", ".join(time_params.keys())
     daysperhour = 1 / 24.
     hourspersec = 1 / 60.**2
     dayspersec = daysperhour * hourspersec
 
-    if dh and not dd:
-        time_params['duration'] = time_params['duration_hours'] * daysperhour
-        dd = True
-    elif dd:
-        time_params['duration'] = time_params['duration_days']
+    if ta:
+        # Time array is defined. Supercedes all other parameters:
+        time_arr = time_params['time_array']
+        time_params['Ntimes'] = len(time_arr)
 
-    if not nt:
-        if not it:
-            raise ValueError("Either integration_time or Ntimes must be "
-                             "included in parameters:" + kws_used)
-        if st and et:
-            time_params['duration'] = time_params['end_time'] - time_params['start_time'] + time_params['integration_time'] * dayspersec
+    else:
+        if dh and not dd:
+            time_params['duration'] = time_params['duration_hours'] * daysperhour
             dd = True
         elif dd:
             time_params['duration'] = time_params['duration_days']
@@ -746,37 +742,30 @@ def parse_time_params(time_params):
     
         inttime_days = time_params['integration_time'] * dayspersec
         if not dd:
-            raise ValueError("Either duration or integration time "
-                             "must be specified: " + kws_used)
-        time_params['integration_time'] = (time_params['duration'] / dayspersec
-                                           / float(time_params['Ntimes']))  # In seconds
-
-    inttime_days = time_params['integration_time'] * dayspersec
-    if not dd:
-        time_params['duration'] = inttime_days * (time_params['Ntimes'])
-        dd = True
-    if not st:
-        if et and dd:
-            time_params['start_time'] = time_params['end_time'] - time_params['duration'] + inttime_days
-    if not et:
-        if st and dd:
-            time_params['end_time'] = time_params['start_time'] + time_params['duration'] - inttime_days
-    if not (st or et):
-        raise ValueError("Either a start or end time must be specified: " + kws_used)
-
-    time_arr = np.linspace(time_params['start_time'],
-                           time_params['end_time'] + inttime_days,
-                           time_params['Ntimes'], endpoint=False)
-
-    if time_params['Ntimes'] != 1:
-        assert np.allclose(np.diff(time_arr), inttime_days * np.ones(time_params["Ntimes"] - 1), atol=dayspersec)   # To nearest second
-
-    return_dict['integration_time'] = (np.ones_like(time_arr, dtype=np.float64)
-                                       * time_params['integration_time'])
+            time_params['duration'] = inttime_days * (time_params['Ntimes'])
+            dd = True
+        if not st:
+            if et and dd:
+                time_params['start_time'] = time_params['end_time'] - time_params['duration'] + inttime_days
+        if not et:
+            if st and dd:
+                time_params['end_time'] = time_params['start_time'] + time_params['duration'] - inttime_days
[…]
