### 2019-03-30 9de7533
Check that unevenly spaced time array works properly with time_array_to_params
