+++ pyuvsim/simsetup.py
@@ -665,7 +665,7 @@ def parse_frequency_params(freq_params):
 
     if freq_params['Nfreqs'] != 1:
         if not np.allclose(np.diff(freq_arr), freq_params['channel_width'] * np.ones(freq_params["Nfreqs"] - 1)):
-            raise ValueError("Frequency array spacings are not equal to channel width."+
+            raise ValueError("Frequency array spacings are not equal to channel width." +
                              "\nInput parameters are: {}".format(str(_freq_params)))
 
     Nspws = 1 if 'Nspws' not in freq_params else freq_params['Nspws']
@@ -718,7 +718,7 @@ def parse_time_params(time_params):
             dd = True
         elif dd:
             time_params['duration'] = time_params['duration_days']
-    
+
         if not nt:
             if not it:
                 raise ValueError("Either integration_time or Ntimes must be "
@@ -732,14 +732,14 @@ def parse_time_params(time_params):
             else:
                 raise ValueError("Either duration or time bounds must be specified: "
                                  + kws_used)
-    
+
         if not it:
             if not dd:
                 raise ValueError("Either duration or integration time "
                                  "must be specified: " + kws_used)
             time_params['integration_time'] = (time_params['duration'] / dayspersec
                                                / float(time_params['Ntimes']))  # In seconds
-    
+
         inttime_days = time_params['integration_time'] * dayspersec
         if not dd:
             time_params['duration'] = inttime_days * (time_params['Ntimes'])
@@ -752,16 +752,16 @@ def parse_time_params(time_params):
                 time_params['end_time'] = time_params['start_time'] + time_params['duration'] - inttime_days
         if not (st or et):
             raise ValueError("Either a start or end time must be specified: " + kws_used)
-    
+
         time_arr = np.linspace(time_params['start_time'],
                                time_params['end_time'] + inttime_days,
                                time_params['Ntimes'], endpoint=False)
-    
+
         if time_params['Ntimes'] != 1:
             if not np.allclose(np.diff(time_arr), inttime_days * np.ones(time_params["Ntimes"] - 1), atol=dayspersec):   # To nearest second
-                raise ValueError("Time array spacings are not equal to integration_time."+
+                raise ValueError("Time array spacings are not equal to integration_time." +
                                  "\nInput parameters are: {}".format(str(_time_params)))
-    
+
         return_dict['integration_time'] = (np.ones_like(time_arr, dtype=np.float64)
                                            * time_params['integration_time'])
     return_dict['time_array'] = time_arr
