### 2019-04-03 a868ae3
Add numerical effects to comparison
