+++ docs/comparison.rst
@@ -2,7 +2,9 @@
 Comparison with other simulators
 ================================
 
-pyuvsim is designed to calculate visibilities from point sources and a given primary beam model. It does this through direct calculation of the full-polarization Radio Interferometer Measurement Equation (RIME), fully parallelized across a set of MPI processes. As such, it has relatively poor performance compared with several other simulators. Deciding which tool to use will ultimately depend on what factors are important to the user. We recommend pyuvsim as a reference tool to verify the results of other simulators, and for producing limited datasets for comparison with observations (particularly on long baselines). Users will need access to a lot of memory and a large number of processor cores to take advantage of pyuvsim's accuracy and precision.
+pyuvsim is designed to calculate visibilities from point sources and a given primary beam model. It does this through direct calculation of the full-polarization Radio Interferometer Measurement Equation (RIME), fully parallelized across a set of MPI processes. Its development has progressed with the goal of providing a tool that can be used reliably and flexibly, but most of all it has focused on avoiding effects that confound foreground modeling efforts in 21cm cosmology.
+
+As such, it has relatively poor performance compared with several other simulators. Deciding which tool to use will ultimately depend on what factors are important to the user. We recommend pyuvsim as a reference tool to verify the results of other simulators, and for producing limited datasets for comparison with observations (particularly on long baselines). Users will need access to a lot of memory and a large number of processor cores to take advantage of pyuvsim's accuracy and precision, ideally on a dedicated high-performance computing cluster.
 
 On this page, we compare pyuvsim's features and performance to several other visibility simulators.
 
@@ -13,7 +15,7 @@ A python package for generating full-sky simulations.
 
 Advantages:
 
-1. A bit faster.
+1. Faster due to vectorized calculations. 
 2. Well-tested through years of use.
 3. Support for a range of sky models including point sources, diffuse emission, and spectral cubes.
 4. Support for tracking as well as transit telescopes.
@@ -25,6 +27,7 @@ Disadvantages:
 1. MPI overhead is large if time axis in simulation is long.
 2. Supports only one polarization at a time.
 3. Does not support non-identical antenna patterns.
+4. Internal phasing methods are in error.
 
 Fast Holographic Deconvolution (FHD)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -48,6 +51,11 @@ Disadvantages:
 5. Primary beam motion on the sky is limited to the "snapshot" size, not the time step size.
 6. Not parallelized over frequency.
 
+.. image:: fhd_uvsim_compare.png
+    :width: 600
+    :alt: Delay-spectrum comparison of FHD and pyuvsim point source simulations.
+
+The figure above shows a comparison of a `pyuvsim` simulation of several point sources with a corresponding FHD simulation. The visibilities have been Fourier-transformed along the frequency axis (a "delay transform") with a Blackman-Harris window function, squared, and then averaged in time. This forms a so-called *delay spectrum*, which is proportional to the power spectrum modes along the line of sight. For smooth-spectrum foregrounds, the power at high delays should be very low. The FHD simulation (shown in orange) introduces excess power at high delays, which can be a real problem for 21cm cosmological research.
 
 Common Astronomy Software Applications (CASA)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -56,11 +64,20 @@ CASA is a set of tools published by the National Radio Astronomical Observatory 
 
 Advantages:
 
-1. Faster, and less
+1. Uses compiled C code internally, which is faster.
 2. Established in the field already and very well-documented.
[…]
