### 2019-04-03 dc49092
Update comparison.rst
Remove an inaccurate comment
