### 2019-04-04 cd99ba4
replace assert with ValueError check
