### 2019-04-04 8b7fe80
Better phrasing and citation on CASA polarization argument
