### 2019-04-04 53f2c08
separate tests of frequency and time parsers
