### 2019-04-04 4e8034c
remove freq/time tests from check_param_reader
