### 2019-04-04 ad35cb4
Remove diagnostic prints
