+++ docs/usage.rst
@@ -12,7 +12,9 @@ Using MPI
 ^^^^^^^^^
 
 ``pyuvsim`` is parallelized using the Message Passing Interface (MPI). To take full advantage of this, any wrapper must be run with ``mpirun``:
-    ::
+
+    .. code-block:: python
+
         # Running with 50 MPI processing units
         > mpirun -n 50 python run_param_pyuvsim obsparam_filename.yaml   # This will run a parameter file job with 10 processing units.
 
