### 2019-04-05 77d8f94
deepcopies in parse_time & freq
