### 2019-04-05 5275075
added setup_uvdata and tests
