+++ pyuvsim/simsetup.py
@@ -39,8 +39,17 @@ def _parse_layout_csv(layout_csv):
         str_format_code = 'a'
     else:
         str_format_code = 'U'
-    dt = np.format_parser([str_format_code + '10', 'i4', 'i4', 'f8', 'f8', 'f8'],
-                          ['name', 'number', 'beamid', 'e', 'n', 'u'], header)
+
+    # get data types for each column
+    dtypes = {"name": str_format_code + '10', 'number': 'i4', 'beamid': 'i4', 'e': 'f8', 'n': 'f8', 'u': 'f8'}
+
+    # check columns in file
+    lower_header = [col.lower() for col in header]
+    columns = ['name', 'number', 'beamid', 'e', 'n', 'u']
+    col_exist = [col for col in columns if col in lower_header]
+
+    dt = np.format_parser([dtypes[col] for col in col_exist],
+                          col_exist, header)
 
     return np.genfromtxt(layout_csv, autostrip=True, skip_header=1,
                          dtype=dt.dtype)
@@ -487,12 +496,15 @@ def initialize_catalog_from_params(obs_params, input_uv=None):
     return np.array(catalog), source_list_name
 
 
-def parse_telescope_params(tele_params, config_path):
+def parse_telescope_params(tele_params, config_path=''):
     """
     Parse the "telescope" section of obsparam.
 
     Args:
         tele_params: Dictionary of telescope parameters
+            See pyuvsim documentation for allowable keys.
+        config_path: path to directory holding configuration and
+            layout files.
 
     Returns:
         dict of array properties:
@@ -502,39 +514,86 @@ def parse_telescope_params(tele_params, config_path):
             |  antenna_numbers: corresponding list of antenna numbers
             |  antenna_positions: Array of ECEF antenna positions
             |  telescope_location: ECEF array center location
+            |  telescope_location_lat_lon_alt: Lat Lon Alt array center location
             |  telescope_config_file: Path to configuration yaml file
             |  antenna_location_file: Path to csv layout file
             |  telescope_name: observatory name
         beam_list:  Beam models in the configuration.
         beam_dict:  Antenna numbers to beam indices
     """
-
-    telescope_config_name = tele_params['telescope_config_name']
-    layout_csv = tele_params['array_layout']
-    if not os.path.isdir(config_path):
-        config_path = os.path.dirname(config_path)
+    tele_params = copy.deepcopy(tele_params)
+    # check for telescope config
+    tele_config = 'telescope_config_name' in tele_params
+    if tele_config:
+        # parse telescope config
         if not os.path.isdir(config_path):
-            raise ValueError('config_path from yaml is not a directory')
-    if not os.path.exists(telescope_config_name):
-        telescope_config_name = os.path.join(config_path, telescope_config_name)
+            config_path = os.path.dirname(config_path)
+            if not os.path.isdir(config_path):
+                raise ValueError('config_path {} is not a directory'.format(config_path))
+        telescope_config_name = tele_params['telescope_config_name']
         if not os.path.exists(telescope_config_name):
-            raise ValueError('telescope_config_name file from yaml does not exist')
+            telescope_config_name = os.path.join(config_path, telescope_config_name)
+            if not os.path.exists(telescope_config_name):
+                raise ValueError('telescope_config_name file from yaml does not exist')
+        with open(telescope_config_name, 'r') as yf:
[…]
