+++ pyuvsim/tests/test_simsetup.py
@@ -223,6 +223,25 @@ def test_param_reader():
         yield (check_param_reader, n)
 
 
+def test_tele_parser():
+    """
+    Check a variety of cases not already tested by param reader
+    """
+    # check no tele config passed
+    tdict = dict(array_layout=os.path.join(SIM_DATA_PATH, 'test_layout_6ant.csv'))
+    nt.assert_raises(KeyError, pyuvsim.simsetup.parse_telescope_params, tdict)
+    tdict['telescope_location'] = '(-30.72152777777791, 21.428305555555557, 1073.0000000093132)'
+    nt.assert_raises(KeyError, pyuvsim.simsetup.parse_telescope_params, tdict)
+    tdict['telescope_name'] = 'tele'
+    tpars, blist, bdict = pyuvsim.simsetup.parse_telescope_params(tdict)
+    nt.assert_equal(tpars['Nants_data'], 6)
+    nt.assert_equal(blist, [])
+    nt.assert_equal(bdict, {})
+
+    tdict.pop('array_layout')
+    nt.assert_raises(KeyError, pyuvsim.simsetup.parse_telescope_params, tdict)
+
+
 def test_freq_parser():
     """
     Check a variety of cases for the frequency parser.
@@ -383,7 +402,7 @@ def test_param_select_cross():
     uv_obj_cross, new_beam_list, new_beam_dict = \
         pyuvsim.initialize_uvdata_from_params(param_dict)
 
-    uv_obj_cross2 = uv_obj_full.select(ant_str='cross', inplace=False, metadata_only=True)
+    uv_obj_cross2 = uv_obj_full.select(ant_str='cross', inplace=False, metadata_only=False)
 
     nt.assert_equal(uv_obj_cross, uv_obj_cross2)
 
@@ -401,7 +420,9 @@ def test_param_select_bls():
     uv_obj_bls, new_beam_list, new_beam_dict = \
         pyuvsim.initialize_uvdata_from_params(param_dict)
 
-    uv_obj_bls2 = uv_obj_full.select(bls=[(40, 41), (42, 43), (44, 45)], inplace=False, metadata_only=True)
+    uv_obj_bls2 = uv_obj_full.select(bls=[(40, 41), (42, 43), (44, 45)], inplace=False, metadata_only=False)
+
+    uv_obj_bls.history, uv_obj_bls2.history = '', ''
 
     nt.assert_equal(uv_obj_bls, uv_obj_bls2)
 
@@ -440,9 +461,59 @@ def test_param_select_redundant():
         pyuvsim.initialize_uvdata_from_params(param_dict)
     nt.assert_true(uv_obj_red.Nbls < uv_obj_full.Nbls)
 
-    uv_obj_red2 = uv_obj_full.compress_by_redundancy(tol=0.1, inplace=False, metadata_only=True)
+    uv_obj_red2 = uv_obj_full.compress_by_redundancy(tol=0.1, inplace=False, metadata_only=False)
+
+    uv_obj_red.history, uv_obj_red2.history = '', ''
 
     nt.assert_equal(uv_obj_red, uv_obj_red2)
+
+
+def test_setup_uvdata():
+    # check it runs through
+    uvd = pyuvsim.simsetup.setup_uvdata(array_layout=os.path.join(SIM_DATA_PATH, "test_config/triangle_bl_layout.csv"),
+                                        telescope_location=(-30.72152777777791, 21.428305555555557, 1073.0000000093132),
+                                        telescope_name="HERA", Nfreqs=10, start_freq=1e8, bandwidth=1e8, Ntimes=60,
+                                        integration_time=100.0, start_time=2458101.0, polarizations=['xx'], no_autos=True,
[…]
