### 2019-04-05 11536af
update setup.py for better installation
