### 2019-04-05 f9ebe76
fix missing comma in setup.py
