### 2019-04-06 6a84daf
pep8 and test fixes
