### 2019-04-06 0428895
pep8 and array_layout dict in setup_uvdata
