+++ pyuvsim/tests/test_mpi_uvsim.py
@@ -58,8 +58,7 @@ def test_run_param_uvsim():
     # Test vot and txt catalogs for parameter simulation
 
     uv_ref = UVData()
-    uvtest.checkWarnings(uv_ref.read_uvfits, [os.path.join(SIM_DATA_PATH, 'testfile_singlesource.uvfits')],
-                         nwarnings=1, message='antenna_diameters is not set')
+    uv_ref.read_uvfits(os.path.join(SIM_DATA_PATH, 'testfile_singlesource.uvfits'))
     uv_ref.unphase_to_drift(use_ant_pos=True)
 
     param_filename = os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testcat.yaml')
@@ -70,9 +69,9 @@ def test_run_param_uvsim():
     tempfilename = params_dict['filing']['outfile_name']
     # This test obsparam file has "single_source.txt" as its catalog.
 
-    uvtest.checkWarnings(pyuvsim.uvsim.run_uvsim, [param_filename], nwarnings=1,
-                         message=['The default for the `center` keyword has changed'],
-                         category=DeprecationWarning)
+    uvtest.checkWarnings(pyuvsim.uvsim.run_uvsim, [param_filename], nwarnings=4,
+                        message=['key'] * 3 + ['The default for the `center` keyword'],
+                        category=[UserWarning]*3 + [DeprecationWarning])
     uv_new_txt = UVData()
     uvtest.checkWarnings(uv_new_txt.read_uvfits, [tempfilename], message='antenna_diameters is not set')
     uv_new_txt.unphase_to_drift(use_ant_pos=True)
@@ -81,19 +80,20 @@ def test_run_param_uvsim():
     param_filename = os.path.join(SIM_DATA_PATH, 'test_config', 'param_1time_1src_testvot.yaml')
 
     uvtest.checkWarnings(pyuvsim.uvsim.run_uvsim, [param_filename],
-                         nwarnings=11,
-                         message=([SIM_DATA_PATH] * 10
+                         nwarnings=14,
+                         message=(['key'] * 3 + [SIM_DATA_PATH] * 10
                                   + ['The default for the `center` keyword has changed']),
-                         category=([astropy.io.votable.exceptions.W50] * 10
+                         category=([UserWarning] * 3 + [astropy.io.votable.exceptions.W50] * 10
                                    + [DeprecationWarning]))
 
     uv_new_vot = UVData()
-    uvtest.checkWarnings(uv_new_vot.read_uvfits, [tempfilename], message='antenna_diameters is not set')
+    uvtest.checkWarnings(uv_new_vot.read_uvfits, [tempfilename], nwarnings=1, category=[UserWarning], message=['antenna_diameters is not set'])
     uv_new_vot.unphase_to_drift(use_ant_pos=True)
     os.remove(tempfilename)
     uv_new_txt.history = uv_ref.history  # History includes irrelevant info for comparison
     uv_new_vot.history = uv_ref.history
     uv_new_txt.object_name = uv_ref.object_name
+    uv_new_vot.object_name = uv_ref.object_name
     nt.assert_equal(uv_new_txt, uv_ref)
     nt.assert_equal(uv_new_vot, uv_ref)
 
