### 2019-04-06 5086cc6
increased test coverage
