### 2019-04-08 0997f9e
comparison updates
