+++ docs/comparison.rst
@@ -4,7 +4,7 @@ Comparison with other simulators
 
 pyuvsim is designed to calculate visibilities from point sources and a given primary beam model. It does this through direct calculation of the full-polarization Radio Interferometer Measurement Equation (RIME), fully parallelized across a set of MPI processes. Its development has progressed with the goal of providing a tool that can be used reliably and flexibly, but most of all it has focused on avoiding effects that confound foreground modeling efforts in 21cm cosmology.
 
-As such, it has relatively poor performance compared with several other simulators. Deciding which tool to use will ultimately depend on what factors are important to the user. We recommend pyuvsim as a reference tool to verify the results of other simulators, and for producing limited datasets for comparison with observations (particularly on long baselines). Users will need access to a lot of memory and a large number of processor cores to take advantage of pyuvsim's accuracy and precision, ideally on a dedicated high-performance computing cluster.
+As such, it has relatively slow speed and large memory usage compared with several other simulators. Deciding which tool to use will ultimately depend on what factors are important to the user. We recommend pyuvsim as a reference tool to verify the results of other simulators, and for producing limited datasets for comparison with observations (particularly on long baselines). Users will need access to a lot of memory and a large number of processor cores to take advantage of pyuvsim's accuracy and precision, ideally on a dedicated high-performance computing cluster.
 
 On this page, we compare pyuvsim's features and performance to several other visibility simulators.
 
@@ -26,6 +26,7 @@ Disadvantages:
 
 1. MPI overhead is large if time axis in simulation is long.
 2. Supports only one polarization at a time.
+3. Limited to accurately simulating the total intensity radio interferometer measurement equation, which is accurate if polarization leakage effects are negligible. This leakage can be due to polarized sky emission or off-diagonal terms in the beam Jones matrix.
 3. Does not support non-identical antenna patterns.
 
 Fast Holographic Deconvolution (FHD)
@@ -35,7 +36,7 @@ FHD is a data calibration and modeling framework written in IDL by researchers a
 
 Advantages:
 
-1. Fast.
+1. Optimized for smooth-spectrum sources.
 2. Well-tested through years of use.
 3. Support for diffuse foreground models.
 4. Support for discrete pointings.
@@ -44,11 +45,9 @@ Advantages:
 Disadvantages:
 
 1. Written in the proprietary IDL language.
-2. Designed for the MWA, and has limited support for other instruments.
-3. The discrete convolution in Fourier Space can introduce aliasing and ringing artifacts at a level relevant to 21cm cosmology.
-4. Measurement equation formalism relies on the flat-sky approximation, limiting curved sky effects.
-5. Primary beam motion on the sky is limited to the "snapshot" size, not the time step size.
-6. Not parallelized over frequency.
+2. The discrete convolution in Fourier Space can introduce aliasing and ringing artifacts at a level relevant to 21cm cosmology.
+3. Primary beam motion on the sky is limited to the "snapshot" size, not the time step size.
+4. Much slower for complex spectral structure.
 
 .. image:: fhd_uvsim_compare.png
     :width: 600
@@ -71,9 +70,10 @@ Advantages:
 Disadvantages:
 
 1. Limited support for user-defined primary beam models.
-2. Lacks full-sky coverage.
-3. Point sources are gridded to pixel locations, losing precision. This pixel-scale imprecision can introduce point source subtraction errors that are significant to 21cm cosmology experiments [CTROTT2012]_.
-4. Full direction-dependent Jones matrices are supported, but simulating with them requires that the beam times sky model calculation be carried out in separate softwa [JAGANNATHAN17]_.
+2. Internal UVW rotation is known to be incorrect, affecting coherence far from the phase center.
[…]
