### 2019-04-08 2e5b552
Update comparison.rst
Reword polarization capability in PRISim
