+++ pyuvsim/source.py
@@ -79,7 +79,7 @@ class SkyModel(object):
 
     _Ncomp_attrs = ['ra', 'dec', 'coherency_radec', 'coherency_local',
                     'stokes', 'alt_az', 'rise_lst', 'set_lst', 'freq',
-                    'pos_lmn', 'name']
+                    'pos_lmn', 'name', 'below_horizon', 'skycoord']
     _scalar_attrs = ['Ncomponents', 'time', 'pos_tol']
 
     _member_funcs = ['coherency_calc', 'update_positions']
@@ -129,32 +129,36 @@ class SkyModel(object):
 
         self.Ncomponents = ra.size
 
+        # If rise/set lsts are not passed in they should be Ncomponent arrays of NaN
+        if rise_lst is None:
+            rise_lst = np.array([np.nan] * self.Ncomponents)
+        if set_lst is None:
+            set_lst = np.array([np.nan] * self.Ncomponents)
+
         self.name = np.asarray(name)
-        self.freq = freq
+        self.freq = np.asarray(freq)
         self.stokes = np.asarray(stokes)
         self.ra = ra
         self.dec = dec
         self.pos_tol = pos_tol
         self.time = None
-        self.rise_lst = rise_lst
-        self.set_lst = set_lst
-
-        # If rise/set lsts are not passed in they should be Ncomponent arrays of NaN
-        if self.rise_lst is None:
-            self.rise_lst = np.array([np.nan] * self.Ncomponents)
-        if self.set_lst is None:
-            self.set_lst = np.array([np.nan] * self.Ncomponents)
+        self.rise_lst = np.asarray(rise_lst)
+        self.set_lst = np.asarray(set_lst)
 
         if self.Ncomponents == 1:
             self.stokes = self.stokes.reshape(4, 1)
             self.ra = self.ra.reshape(1, 1)
             self.dec = self.dec.reshape(1, 1)
             self.name = self.name.reshape(1, 1)
+            self.rise_lst = self.rise_lst.reshape(1, 1)
+            self.set_lst = self.set_lst.reshape(1, 1)
 
         assert np.all([self.Ncomponents == l for l in
                        [self.ra.size, self.dec.size, self.freq.size, self.stokes.shape[1]]]), 'Inconsistent quantity dimensions.'
 
         self.skycoord = SkyCoord(self.ra, self.dec, frame='icrs')
+
+        self.below_horizon = np.zeros(self.Ncomponents).astype(bool)
 
         # The coherency is a 2x2 matrix giving electric field correlation in Jy.
         # Multiply by .5 to ensure that Trace sums to I not 2*I
@@ -163,6 +167,13 @@ class SkyModel(object):
                                                self.stokes[2, :] - 1j * self.stokes[3, :]],
                                               [self.stokes[2, :] + 1j * self.stokes[3, :],
                                                self.stokes[0, :] - self.stokes[1, :]]])
+
+    def split(self, N):
+        """
+        Split this into a list of N SkyModel objects
+
+        Numbers of components per object follows the splitting logic of numpy.array_split
+        """
 
     def select(self, s):
         """
@@ -184,7 +195,8 @@ class SkyModel(object):
 
     def __getitem__(self, i):
         # i = valid index object
-        return _view_source(self, i)
+        self._select = i
+        return self
 
     def __iter__(self):
         self._counter = 0
@@ -205,7 +217,7 @@ class SkyModel(object):
          # For python 2 compatibility
         return self.__next__()
 
-    def coherency_calc(self, time, telescope_location, src_inds=slice(None)):
+    def coherency_calc(self, telescope_location, src_inds=slice(None)):
         """
         Calculate the local coherency in alt/az basis for this source at a time & location.
 
@@ -214,17 +226,12 @@ class SkyModel(object):
[…]
