### 2019-04-12 fe2e37b
clarify mpi4py installation instructions
