### 2019-04-12 7d56da3
update pyuvdata/numpy/mpi4py minimum versions
