+++ requirements.txt
@@ -1,6 +1,6 @@
-pyuvdata==1.2.1
+pyuvdata>=1.3.7
 six==1.11.0
-mpi4py==3.0.1a0
+mpi4py>=3.0.0
 astropy==2.0.6
 setuptools==38.4.0
 numpy==1.11.3
