### 2019-04-12 b32223b
update comparison docs
