### 2019-04-12 cbf1d04
pstokes power beam in analyticbeam.py
