### 2019-04-12 432df32
start power law gaussian beam widths
