+++ requirements.txt
@@ -1,9 +1,7 @@
-pyuvdata>=1.3.7
-six==1.11.0
+numpy>=1.15
+six>=1.11.0
+setuptools>=38.4.0
+scipy>=1.0.1
+PyYAML>=5.1
 mpi4py>=3.0.0
-astropy==2.0.6
-setuptools==38.4.0
-numpy==1.11.3
-nose==1.3.7
-scipy==1.0.1
-PyYAML==5.1
+astropy>=2.0.6
