### 2019-04-12 e9c796d
test chromatic gaussian widths
