### 2019-04-13 8935550
update repo installation instructions
