+++ README.md
@@ -19,7 +19,11 @@ The two primary pyuvsim goals are interferometer simulation accuracy at the leve
 
 # Installation
 * For simple installation, the latest stable release is available via pip (`pip install pyuvsim`)
-* To install the development version: Clone the repository using `git clone https://github.com/RadioAstronomySoftwareGroup/pyuvsim`, navigate into the pyuvsim directory and run `pip install .`
+* To install the development version: Clone the repository using `git clone https://github.com/RadioAstronomySoftwareGroup/pyuvsim`, navigate into the pyuvsim directory and run
+```
+pip install -r requirements.txt
+pip install .
+```
 * `pyuvsim` is mainly intended to run on clusters running the linux operating system.
 --
 **Note**
