### 2019-04-13 49a115b
test power beam definition for gaussian beams
