### 2019-04-17 5286029
rephrasing, and adding a point about standards
