### 2019-04-17 37b3dd0
rephrasing, and adding a point about standards
