### 2019-04-17 cecff7f
Small wording improvements.
