### 2019-04-17 431dd5f
fix numbering in comparison.rst
