### 2019-04-17 9c8ca5f
redo Note block in README
