### 2019-04-17 #1234 @aelanman
Very sorry for the delay! I believe we've addressed all of the comments with the latest PR.

> **Paper**
>     * 1 minor comment: you list future low-frequency telescopes. Any reason you don't list LOFAR which is a current low-frequency radio telescope? I think such a tool could be very useful for current telescopes as well (for algorithm development and other R&D on the array).

Major oversight on my part! It's included now.
>
>     * In the readthedocs, you give a nice list of other tools that simulate radio data. You then mention what sets you apart from them (i.e. they all use apporximations to speed up). I would add this to the paper and if possible, would it be possible to quantify the effect of your fully analytic approach on the performance and precision (do I gain 10% precision in 200% of the run time; these kind of numbers)? This is the only comment that would require a bit of extra work. I am not really sure what the policy of JOSS is in this respect and whether this is in the category "nice to have" or "essential" in that case (@arfon: any comments on this?).

As per our previous discussion, I've included a new page linked on the RTD page that compares features of pyuvsim with three other simulators used in our field. For FHD I've briefly discussed the high delay power it introduces, when run as a simulator, as an example of one numerical artifact that pyuvsim avoids. For CASA I mention that its most common mode of operation grids point sources before calculating visibilities, which can introduce source modeling errors. The main context of the precision and accuracy improvements is that we need to avoid introducing spectral structure in foreground source power, and both of these effects are significant for our field of 21cm cosmology. I also mention that although it can support for polarized source models and wide fields of view, these uses are limited to customized wrappers tools and user-defined tools and are not the default operation.

>     * Also it would be nice if either the paper or the documentation could give potentially interested users a rough idea whether using this tool is the best fit to their use case as opposed to CASA, OSCAR, FHD, PRISim, etc.

The opening paragraph of the new `comparison` document cautions the user about the performance limitations of pyuvsim, as well as its robustness and applicability.

>
>     * "The measurement equation": add citation (Hamaker et al. 1996 ).
It's cited now. Thanks for pointing that out!
>
>
> **README.md**
>
>     * The README has a section "Inputs" but doesn't seem to mention outputs (It's a standard uvfits I guess so it would be a single line)

A brief output section is included. pyuvsim is capable of writing out to any file format compatible with pyuvdata, so most of that we leave to the pyuvdata documentation.
>
>     * In both the README and the documentation the following quick-run command is given:
>       `mpirun -n 20 python run_param_pyuvsim.py reference_simulations/obsparam_1.1.yaml`
>       It should be
>       `mpirun -n 20 python scripts/run_param_pyuvsim.py reference_simulations/obsparam_ref_1.1.yaml`
>       and specify where to execute it from
>
Fixed.

>     * Community guidelines are present, but it doesn't really say where one can get support (questions that are not per se bug reports). Is that via github issues?
I've added a little bit to the README clarify this, but support questions should be given through github issues.
>
>
> **readthedocs documentation**
>
>     * "Parameter and configuration Files" section doesn't state where all the files should be. I assume the paths are relative to the path of the main yaml file but It's best to explicitly write that down.

I've added a line to the file to clarify that the path can either be absolute or specified relative to the location of the obsparam file.

>     * In the "Enabling Profiling" section
>       `To run a simulation with profiling enabled, run the command profiling.set_profiler before starting`
>       change to`(...) run the command profiling.set_profiler() before `
Fixed.
>
>
> **installation instruction and procedure**
>
>     * Installation: you only present how to build the master branch. Since you have releases, I think you should explicitly document hoe to build a (the latest?) release.
Installation instructions include a `pip` option, which will install the latest stable release. We will issue a new release when this review is finished as well.
>
>     * I would strongly advice to copy the installation instructions from readthedocs to the README.md file too.

The landing page of RTD is generated from the README.md file, so the installation instructions are on both.
>
>     * Adding a requirements.txt could also make sense although I did notice that pyuvsim doesn't really do requirement installation well.

We've added a requirements.txt and set minimum versions on all requirements. The installation instructions have been updated in the readme as well. We're also working on an anaconda package which should make automated installation much easier, but this will be built with the next release.

>     * Also listing max an min versions if applicable (seems mpi4py 1.3.1 is too old for example; this is the version from Centos7 EPEL).
As mentioned, we've set minimum version requirements. `mpi4py` needs to be at least version 3.0.
>
> **tests**
>
>     * Is there a reason you don't add the availability of nosetests to the documentation? If I know a package is testable, I will test it especially if it documents how to test it.
It's mentioned now.
>
>     * One test fails if h5py is not installed. Probably it is a dependency via pyuvdata, but it's worth documenting that you need h5py to pass all tests.
We've added "test_requires" to setup.py, which includes h5py.
>
>     * Also if you run the tests twice, the `test_file.uvfits` file will already exist and the system will cope by creating `test_file_0.uvfits` making the corresponding test to fail.

I'm unable to recreate this behavior. Is it possible that one of your tests was interrupted before it could clean up?
