### 2019-04-22 66c8abf
call assert_raises correctly
