### 2019-04-24 e4a7cdb
construct power beam from efield as done in UVBeam
