### 2019-04-25 9eed3ae
set off-diagonal terms in analytic jones matrices to zero
