+++ pyuvsim/antenna.py
@@ -45,8 +45,8 @@ class Antenna(object):
 
         # convert to UVBeam az/za convention
         source_za, source_az = simutils.altaz_to_zenithangle_azimuth(source_alt_az[0], source_alt_az[1])
-        source_za = np.array([source_za])
-        source_az = np.array([source_az])
+        source_za = np.asarray([source_za])
+        source_az = np.asarray([source_az])
 
         freq = np.array([frequency.to('Hz').value])
 
@@ -59,14 +59,15 @@ class Antenna(object):
                                                  za_array=source_za,
                                                  freq_array=freq,
                                                  reuse_spline=reuse_spline)
+        Ncomponents = source_za.shape[-1]
 
-        # interp_data has shape: (Naxes_vec, Nspws, Nfeeds, 1 (freq), 1 (source position))
-        jones_matrix = np.zeros((2, 2), dtype=np.complex)
+        # interp_data has shape: (Naxes_vec, Nspws, Nfeeds, 1 (freq),  Ncomponents (source positions))
+        jones_matrix = np.zeros((2, 2, Ncomponents), dtype=np.complex)
         # first axis is feed, second axis is theta, phi (opposite order of beam!)
-        jones_matrix[0, 0] = interp_data[1, 0, 0, 0, 0]
-        jones_matrix[1, 1] = interp_data[0, 0, 1, 0, 0]
-        jones_matrix[0, 1] = interp_data[0, 0, 0, 0, 0]
-        jones_matrix[1, 0] = interp_data[1, 0, 1, 0, 0]
+        jones_matrix[0, 0] = interp_data[1, 0, 0, 0, :]
+        jones_matrix[1, 1] = interp_data[0, 0, 1, 0, :]
+        jones_matrix[0, 1] = interp_data[0, 0, 0, 0, :]
+        jones_matrix[1, 0] = interp_data[1, 0, 1, 0, :]
 
         return jones_matrix
 
