### 2019-04-25 c2663aa
sum over source components within visibility calculation
