+++ pyuvsim/tests/test_uvsim.py
@@ -33,33 +33,38 @@ longbl_uvfits_file = os.path.join(SIM_DATA_PATH, '5km_triangle_1time_1chan.uvfit
 
 def test_visibility_single_zenith_source():
     """Test single zenith source."""
-    time = Time('2018-03-01 00:00:00', scale='utc')
 
+    beam0 = simtest.make_cst_beams()
+    beam1 = pyuvsim.AnalyticBeam('uniform')
+    beam2 = pyuvsim.AnalyticBeam('gaussian', sigma=np.radians(10.0))
+    beam3 = pyuvsim.AnalyticBeam('airy', diameter=14.0)
+
+    time = Time('2018-03-01 00:00:00', scale='utc')
+    
     array_location = EarthLocation(lat='-30d43m17.5s', lon='21d25m41.9s',
                                    height=1073.)
     time.location = array_location
-
+    
     freq = (150e6 * units.Hz)
     source_arr, _ = pyuvsim.create_mock_catalog(time, arrangement='zenith')
     source = source_arr[0]
     source.update_positions(time, array_location)
-
+    
     antenna1 = pyuvsim.Antenna('ant1', 1, np.array([0, 0, 0]), 0)
     antenna2 = pyuvsim.Antenna('ant2', 2, np.array([107, 0, 0]), 0)
-
+    
     baseline = pyuvsim.Baseline(antenna1, antenna2)
 
-    beam = simtest.make_cst_beams()
-
-    beam_list = [beam]
-    array = pyuvsim.Telescope('telescope_name', array_location, beam_list)
-
-    task = pyuvsim.UVTask(source, time, freq, baseline, array)
-
-    engine = pyuvsim.UVEngine(task)
-
-    visibility = engine.make_visibility()
-    nt.assert_true(np.allclose(visibility, np.array([.5, .5, 0, 0]), atol=5e-3))
+    for beam in [beam0, beam1, beam2, beam3]:
+        beam_list = [beam]
+        array = pyuvsim.Telescope('telescope_name', array_location, beam_list)
+    
+        task = pyuvsim.UVTask(source, time, freq, baseline, array)
+    
+        engine = pyuvsim.UVEngine(task)
+    
+        visibility = engine.make_visibility()
+        nt.assert_true(np.allclose(visibility, np.array([.5, .5, 0, 0]), atol=5e-3))
 
 
 def test_visibility_source_below_horizon():
@@ -398,7 +403,7 @@ def test_file_to_tasks():
     hera_uv = UVData()
     hera_uv.read_uvfits(EW_uvfits_file)
     time = Time(hera_uv.time_array[0], scale='utc', format='jd')
-    sources, _ = pyuvsim.create_mock_catalog(time, arrangement='zenith')
+    sources, _ = pyuvsim.create_mock_catalog(time, arrangement='zenith', Nsrcs=5)
 
     beam = simtest.make_cst_beams()
     beam_list = [beam]
@@ -408,8 +413,9 @@ def test_file_to_tasks():
     Ntimes = hera_uv.Ntimes
     Nfreqs = hera_uv.Nfreqs
     Nsrcs = len(sources)
+    print(Nsrcs)
 
-    Ntasks = Nblts * Nfreqs * Nsrcs
+    Ntasks = Nblts * Nfreqs# * Nsrcs
     beam_dict = None
 
     taskiter = pyuvsim.uvdata_to_task_iter(np.arange(Ntasks), hera_uv, sources, beam_list, beam_dict)
@@ -475,7 +481,6 @@ def test_file_to_tasks():
         exp_task = expected_task_list[idx]
         nt.assert_equal(task, exp_task)
 
-test_file_to_tasks()
 
 def test_uvdata_init():
     hera_uv = UVData()
