### 2019-04-26 423c4bb
handle cases of only one source
