+++ pyuvsim/antenna.py
@@ -45,15 +45,12 @@ class Antenna(object):
 
         # convert to UVBeam az/za convention
         source_za, source_az = simutils.altaz_to_zenithangle_azimuth(source_alt_az[0], source_alt_az[1])
-        source_za = np.asarray([source_za])
-        source_az = np.asarray([source_az])
 
         freq = np.array([frequency.to('Hz').value])
 
         if array.beam_list[self.beam_id].data_normalization != 'peak':
             array.beam_list[self.beam_id].peak_normalize()
         array.beam_list[self.beam_id].interpolation_function = 'az_za_simple'
-
         interp_data, interp_basis_vector = \
             array.beam_list[self.beam_id].interp(az_array=source_az,
                                                  za_array=source_za,
