### 2019-04-26 fb26bdb
remove sources from task iterator
