### 2019-04-26 #1234 @ygrange
Thanks for the complete reply to all the points I made! I will not reply to each of them separately because I agree with all points you make there.

Nice to see that testing is not only mentioned in the README, but decent testing is part of the design to begin with.

Paper looks good!

The following comments al are rather minor.
* It may be sensible to link to comparison.rst from README.rst but that's fully up to you.

Installing on a clean Docker container: Seems like pyuvdata is not in the requirements.txt. The main other issue I encountered is that mpi4py depends on mpi (yeah, that indeed is quite obvious in hindsight ;)). I don't really consider that your problem though it may actually be fixed in the anaconda build, which is a great idea of you to have.

Please note that when I run the tests in a Docker container running ubuntu 18.04 on my mac laptop (2 cores), I will get failing tests:
```
pyuvsim.tests.test_mpi_uvsim.test_run_uvsim ... ERROR
pyuvsim.tests.test_mpi_uvsim.test_run_param_uvsim ... antenna_diameters is not set. Using known values for HERA.
[f58409b099d2:05741] *** Process received signal ***
[f58409b099d2:05741] Signal: Floating point exception (8)
[f58409b099d2:05741] Signal code: Integer divide-by-zero (1)
[f58409b099d2:05741] Failing at address: 0x7fa471a760a0
[f58409b099d2:05741] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x3ef20)[0x7fa475204f20]
[f58409b099d2:05741] [ 1] /usr/lib/x86_64-linux-gnu/libopen-pal.so.20(+0x820a0)[0x7fa471a760a0]
[f58409b099d2:05741] [ 2] /usr/lib/x86_64-linux-gnu/libopen-pal.so.20(opal_progress_set_event_poll_rate+0x17)[0x7fa471a1ab47]
[f58409b099d2:05741] [ 3] /usr/lib/x86_64-linux-gnu/libopen-pal.so.20(opal_progress_init+0x19)[0x7fa471a1ab89]
[f58409b099d2:05741] [ 4] /usr/lib/x86_64-linux-gnu/libopen-pal.so.20(opal_init+0x151)[0x7fa471a1bc31]
[f58409b099d2:05741] [ 5] /usr/lib/x86_64-linux-gnu/libopen-rte.so.20(orte_init+0xc9)[0x7fa471cba079]
[f58409b099d2:05741] [ 6] /usr/lib/x86_64-linux-gnu/libmpi.so.20(ompi_mpi_init+0x30e)[0x7fa471f7527e]
[f58409b099d2:05741] [ 7] /usr/lib/x86_64-linux-gnu/libmpi.so.20(MPI_Init+0xb9)[0x7fa471f962f9]
[f58409b099d2:05741] [ 8] /usr/local/lib/python2.7/dist-packages/mpi4py/MPI.so(+0x87218)[0x7fa4722a7218]
[f58409b099d2:05741] [ 9] /usr/bin/python(PyEval_EvalFrameEx+0x54a)[0x564287aae4ca]
[f58409b099d2:05741] [10] /usr/bin/python(PyEval_EvalCodeEx+0x6da)[0x564287aabd0a]
[f58409b099d2:05741] [11] /usr/bin/python(PyEval_EvalFrameEx+0x5cb8)[0x564287ab3c38]
[f58409b099d2:05741] [12] /usr/bin/python(PyEval_EvalCodeEx+0x6da)[0x564287aabd0a]
[f58409b099d2:05741] [13] /usr/bin/python(PyEval_EvalFrameEx+0x567e)[0x564287ab35fe]
[f58409b099d2:05741] [14] /usr/bin/python(PyEval_EvalCodeEx+0x6da)[0x564287aabd0a]
[f58409b099d2:05741] [15] /usr/bin/python(+0x10f619)[0x564287ac7619]
[f58409b099d2:05741] [16] /usr/bin/python(PyObject_Call+0x3e)[0x564287a9777e]
[f58409b099d2:05741] [17] /usr/bin/python(PyEval_EvalFrameEx+0x2aa1)[0x564287ab0a21]
[f58409b099d2:05741] [18] /usr/bin/python(PyEval_EvalFrameEx+0x52b2)[0x564287ab3232]
[f58409b099d2:05741] [19] /usr/bin/python(PyEval_EvalCodeEx+0x6da)[0x564287aabd0a]
[f58409b099d2:05741] [20] /usr/bin/python(+0x10f8bc)[0x564287ac78bc]
[f58409b099d2:05741] [21] /usr/bin/python(PyObject_Call+0x3e)[0x564287a9777e]
[f58409b099d2:05741] [22] /usr/bin/python(PyEval_EvalFrameEx+0x2aa1)[0x564287ab0a21]
[f58409b099d2:05741] [23] /usr/bin/python(PyEval_EvalCodeEx+0x6da)[0x564287aabd0a]
[f58409b099d2:05741] [24] /usr/bin/python(+0x10f619)[0x564287ac7619]
[f58409b099d2:05741] [25] /usr/bin/python(+0x1280de)[0x564287ae00de]
[f58409b099d2:05741] [26] /usr/bin/python(PyObject_Call+0x3e)[0x564287a9777e]
[f58409b099d2:05741] [27] /usr/bin/python(+0x1845f7)[0x564287b3c5f7]
[f58409b099d2:05741] [28] /usr/bin/python(PyEval_EvalFrameEx+0x54a0)[0x564287ab3420]
[f58409b099d2:05741] [29] /usr/bin/python(PyEval_EvalFrameEx+0x52b2)[0x564287ab3232]
[f58409b099d2:05741] *** End of error message ***
Floating point exception
```
However I am not fully convinced it is a bug on your side because I cannot reproduce it inside a docker container running on a centos7 (64-core) system. So in that case: be aware that users may report this behaviour.

If I test a checkout from master, two tests fail:
pyuvsim.tests.test_version.test_construct_version_info (Fail)  (I think that's because "master" != "version 0.2.3")
"Test function that defines filenames from parameter dict" (Error because of dependency on h5py)

Would it be possible to make those tests conditional in somw way? Especially the one that depends on h5py.
