### 2019-04-30 3465ca5
add shared memory array option to mpi.py
