### 2019-04-30 ec18c0f
separate flux/horizon cuts from array_to_skymodel function
