+++ pyuvsim/simsetup.py
@@ -91,20 +91,43 @@ def skymodel_to_array(sky):
     arr['source_id'] = sky.name
     arr['ra_j2000'] = sky.ra.value
     arr['dec_j2000'] = sky.dec.value
-    arr['flux_density_I'] = sky.stokes[0,:]
+    arr['flux_density_I'] = sky.stokes[0, :]
     arr['frequency'] = sky.freq
 
     return arr
 
 
-
-def array_to_skymodel(catalog_table, lst_array=None, latitude_deg=None, horizon_buffer=0.04364, min_flux=None, max_flux=None):
+def array_to_skymodel(catalog_table):
     """
-    Generate a SkyModel object from a recarray, performing flux and horizon selections.
+    Make a SkyModel object from a recarray.
+
+    """
+
+    ra = Angle(catalog_table['ra_j2000'], units.deg)
+    dec = Angle(catalog_table['dec_j2000'], units.deg)
+    ids = catalog_table['source_id']
+    freqs = catalog_table['frequency'] * units.Hz
+    flux_I = np.atleast_1d(catalog_table['flux_density_I'])
+    stokes = np.pad(np.expand_dims(flux_I, 1), ((0, 0), (0, 3)), 'constant').T
+    rise_lst = None
+    set_lst = None
+    if 'rise_lst' in catalog_table.dtype.names:
+        rise_lst = catalog_table['rise_lst']
+        set_lst = catalog_table['set_lst']
+    sourcelist = SkyModel(ids, ra, dec, freqs, stokes, rise_lst=rise_lst, set_lst=set_lst)
+
+    return sourcelist
+
+
+def source_cuts(catalog_table, input_uv=None, lst_array=None, latitude_deg=None, horizon_buffer=0.04364, min_flux=None, max_flux=None):
+    """
+    Performing flux and horizon selections on recarray of source components.
 
     Args:
         catalog_table: recarray of source catalog information. Must have columns with names
                        'source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency'
+        input_uv: (UVData)
+            If not None, can set lst_array and latitude_deg if either or both is missing.
         lst_array: For coarse RA horizon cuts, lsts used in the simulation [radians]
         latitude_deg: Latitude of telescope in degrees. Used for declination coarse horizon cut.
         horizon_buffer: Angle (float, in radians) of buffer for coarse horizon cut. Default is about 10 minutes of sky rotation.
@@ -117,7 +140,17 @@ def array_to_skymodel(catalog_table, lst_array=None, latitude_deg=None, horizon_
 
         min_flux: Minimum stokes I flux to select [Jy]
         max_flux: Maximum stokes I flux to select [Jy]
+    Returns:
+        A new recarray of source components, with additional columns for rise and set lst.
+
     """
+
+    # Option to pass a UVData object to get telescope-specific parameters
+    if input_uv:
+        if lst_array is None:
+            lst_array, inds = np.unique(input_uv.lst_array, return_index=True)
+        if latitude_deg is None:
+            latitude_deg = input_uv.telescope_location_lat_lon_alt_degrees[0]
 
     sourcelist = []
     if len(catalog_table.shape) == 0:
@@ -125,6 +158,7 @@ def array_to_skymodel(catalog_table, lst_array=None, latitude_deg=None, horizon_
     Nsrcs = catalog_table.shape[0]
     coarse_kwds = [lst_array, latitude_deg]
     coarse_horizon_cut = all([k is not None for k in coarse_kwds])
+
     if (not coarse_horizon_cut) and any([k is not None for k in coarse_kwds]):
[…]
