+++ pyuvsim/tests/test_analyticbeam.py
@@ -30,37 +30,25 @@ def test_uniform_beam():
     time = Time('2018-03-01 00:00:00', scale='utc')
     array_location = EarthLocation(lat='-30d43m17.5s', lon='21d25m41.9s',
                                    height=1073.)
-    source_list, mock_keywords = pyuvsim.create_mock_catalog(time, 'hera_text', array_location=array_location)
+    sources, mock_keywords = pyuvsim.create_mock_catalog(time, 'hera_text', array_location=array_location)
 
-    nsrcs = len(source_list)
+    nsrcs = len(sources)
     az_vals = []
     za_vals = []
     freq_vals = []
-    for src in source_list:
-        src_coord = SkyCoord(ra=src.ra, dec=src.dec, frame='icrs', obstime=time,
-                             location=array_location)
-        src_coord_altaz = src_coord.transform_to('altaz')
 
-        beam_za, beam_az = simutils.altaz_to_zenithangle_azimuth(src_coord_altaz.alt.to('rad').value,
-                                                                 src_coord_altaz.az.to('rad').value)
-        az_vals.append(beam_az)
-        za_vals.append(beam_za)
-
-        if len(freq_vals) > 0:
-            if src.freq.to('Hz').value != freq_vals[0]:
-                freq_vals.append(src.freq.to('Hz').value)
-        else:
-            freq_vals.append(src.freq.to('Hz').value)
+    sources.update_positions(time, array_location)
+    za_vals = np.pi / 2. - sources.alt_az[1]  # rad
+    az_vals = sources.alt_az[1]
+    freq_vals = sources.freq
 
     n_freqs = len(freq_vals)
-    interpolated_beam, interp_basis_vector = beam.interp(az_array=np.array(az_vals),
-                                                         za_array=np.array(za_vals),
-                                                         freq_array=np.array(freq_vals))
+    interpolated_beam, interp_basis_vector = beam.interp(az_array=az_vals,
+                                                         za_array=za_vals,
+                                                         freq_array=freq_vals)
     expected_data = np.zeros((2, 1, 2, n_freqs, nsrcs), dtype=np.float)
     expected_data[1, 0, 0, :, :] = 1
     expected_data[0, 0, 1, :, :] = 1
-    expected_data[1, 0, 1, :, :] = 1
-    expected_data[0, 0, 0, :, :] = 1
     nt.assert_true(np.allclose(interpolated_beam, expected_data))
 
 
@@ -73,49 +61,34 @@ def test_airy_beam_values():
     time = Time('2018-03-01 00:00:00', scale='utc')
     array_location = EarthLocation(lat='-30d43m17.5s', lon='21d25m41.9s',
                                    height=1073.)
-    source_list, mock_keywords = pyuvsim.create_mock_catalog(time, 'hera_text', array_location=array_location)
+    sources, mock_keywords = pyuvsim.create_mock_catalog(time, 'hera_text', array_location=array_location)
 
-    nsrcs = len(source_list)
+    nsrcs = len(sources)
     az_vals = []
     za_vals = []
     freq_vals = []
-    for src in source_list:
-        src_coord = SkyCoord(ra=src.ra, dec=src.dec, frame='icrs', obstime=time,
-                             location=array_location)
-        src_coord_altaz = src_coord.transform_to('altaz')
 
-        beam_za, beam_az = simutils.altaz_to_zenithangle_azimuth(src_coord_altaz.alt.to('rad').value,
[…]
