+++ pyuvsim/uvsim.py
@@ -98,6 +98,10 @@ class UVEngine(object):
         """ Get apparent coherency from jones matrices and source coherency. """
         baseline = self.task.baseline
         sources = self.task.sources
+
+        if sources.alt_az is None:
+            sources.update_positions(self.task.time, self.task.telescope.location)
+
         # coherency is a 2x2 matrix
         # [ |Ex|^2, Ex* Ey, Ey* Ex |Ey|^2 ]
         # where x and y vectors along the local alt/az axes.
@@ -142,7 +146,8 @@ class UVEngine(object):
         return vis_vector
 
 
-def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict):
+def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict,
+                        tasks_shape=None, axis_inds=None):
     """
     Generate local tasks, reusing quantities where possible.
 
@@ -150,14 +155,13 @@ def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict):
         task_ids (numpy.ndarray of ints): Task index in the full flattened meshgrid of parameters.
         input_uv (UVData): UVData object to use
         catalog: a SkyModel object
-        beam_list: (list of UVBeam or AnalyticBeam objects
-        beam_dict (dict, optional): dict mapping antenna number to beam index in beam_list
-
+        beam_list: list of UVBeam or AnalyticBeam objects
+        beam_dict: dict mapping antenna number to beam index in beam_list
+        tasks_shape: Shape of the unraveled task array.
+        axis_inds: list giving the index of (baseline, time, freqs, source) in tasks_shape
     Yields:
         Iterable of task objects to be done on current rank.
     """
-
-    # Loops, outer to inner: times, sources, frequencies, baselines
     # The task_ids refer to tasks on the flattened meshgrid.
     if not isinstance(input_uv, UVData):
         raise TypeError("input_uv must be UVData object")
@@ -182,6 +186,14 @@ def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict):
     Nsrcs = catalog.Ncomponents
     Nbls = input_uv.Nbls
 
+    if axis_inds is None:
+        axis_inds = range(4)
+        if not tasks_shape is None:
+            raise ValueError("tasks_shape and axis_inds must both be provided")
+        tasks_shape = (Nbls, Ntimes, Nfreqs, Nsrcs)
+
+    bl_ax, time_ax, freq_ax, src_ax = axis_inds
+
     telescope = Telescope(input_uv.telescope_name,
                           EarthLocation.from_geocentric(*input_uv.telescope_location, unit='m'),
                           beam_list)
@@ -189,9 +201,18 @@ def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict):
     freq_array = input_uv.freq_array * units.Hz
     time_array = Time(input_uv.time_array, scale='utc', format='jd', location=telescope.location)
 
+    prev_bltfi = None
     # Shape indicates slowest to fastest index. (time is slowest, baselines is fastest).
     for task_index in task_ids:
-        time_i, freq_i, bl_i = np.unravel_index(task_index, (Ntimes, Nfreqs, Nbls))
+        unrav_ind = np.unravel_index(task_index, tasks_shape)
+        bl_i = unrav_ind[bl_ax]
+        time_i = unrav_ind[time_ax]
+        freq_i = unrav_ind[freq_ax]
+        src_i = unrav_ind[src_ax]
+        if prev_bltfi == (bl_i, time_i, freq_i):
+            continue
[…]

[…]
