### 2019-05-02 #1234 @aelanman
@arfon @ygrange Thank you for looking over the revisions. I've been trying to reproduce your errors. Are you installing via the repo or through the PyPi package?

The h5py dependence is only in one part of one test and should be easy enough to remove, so I'll do that.

I will also put a link to the comparison page in the README.
