### 2019-05-03 c7da438
Add test of shared memory array
