### 2019-05-03 b0f652a
speed up redundancy test
