### 2019-05-03 02b8ad8
Remove source axis from main task loop
