### 2019-05-05 8bdcadb
estimate size of skymodel object in memory
