### 2019-05-05 d60d9b5
utility function to estimate available system memory
