+++ .travis.yml
@@ -29,7 +29,7 @@ install:
   - conda info -a
 
   # install dependencies
-  - conda install numpy scipy astropy six nose pip pyyaml line_profiler pycodestyle coveralls h5py
+  - conda install numpy scipy astropy six nose pip pyyaml line_profiler pycodestyle coveralls h5py psutil
   - conda install -c conda-forge mpi4py
   - pip install git+https://github.com/RadioAstronomySoftwareGroup/pyuvdata.git
   - python setup.py install
