+++ pyuvsim/uvsim.py
@@ -146,6 +146,37 @@ class UVEngine(object):
         return vis_vector
 
 
+def _make_task_inds(Nbls, Ntimes, Nfreqs, Nsrcs, rank, Npus):
+    """
+    Make iterators defining task and sources computed on rank.
+
+       Three options ---
+           (1) Npus < Nbltf -- Split by Nbltf, split sources in the task loop for memory's sake.
+           (2) Nbltf < Npus and Nsrcs > Npus -- Split by Nsrcs only
+           (3) (Nsrcs, Nbltf) < Npus -- Split by Nbltf
+       - Split by instrument axes here.
+       - Within the task loop, decide on source chunks and make skymodels on the fly.
+    """
+
+    Nbltf = Nbls * Ntimes * Nfreqs
+
+    split_srcs = False
+
+    if (Nbltf < Npus) and (Npus < Nsrcs):
+        split_srcs = True
+
+    if split_srcs:
+        src_inds, Nsrcs_local = simutils.iter_array_split(rank, Nsrcs, Npus)
+        task_inds = six.moves.range(Nbltf)
+        Ntasks_local = Nbltf
+    else:
+        task_inds, Ntasks_local = simutils.iter_array_split(rank, Nbltf, Npus)
+        src_inds = six.moves.range(Nsrcs)
+        Nsrcs_local = Nsrcs
+
+    return task_inds, src_inds, Ntasks_local, Nsrcs_local
+
+
 def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict,
                         tasks_shape=None, axis_inds=None):
     """
@@ -154,11 +185,11 @@ def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict,
     Args:
         task_ids: (range iterator) Task index in the full flattened meshgrid of parameters.
         input_uv (UVData): UVData object to use
-        catalog: a SkyModel object
+        catalog: a recarray of source components
         beam_list: list of UVBeam or AnalyticBeam objects
         beam_dict: dict mapping antenna number to beam index in beam_list
         tasks_shape: Shape of the unraveled task array.
-        axis_inds: list giving the index of (baseline, time, freqs, source) in tasks_shape
+        axis_inds: list giving the index of (time, freqs, baseline) in tasks_shape
     Yields:
         Iterable of task objects to be done on current rank.
     """
@@ -166,10 +197,26 @@ def uvdata_to_task_iter(task_ids, input_uv, catalog, beam_list, beam_dict,
     if not isinstance(input_uv, UVData):
         raise TypeError("input_uv must be UVData object")
 
-    if not isinstance(catalog, SkyModel):
-        raise TypeError("catalog must be a SkyModel object")
+#   Skymodel will now be passed in as a catalog array.
+    if not isinstance(catalog, np.ndarray):
+        raise TypeError("catalog must be a record array")
 
-    # There will always be relatively few antennas, so just build the full list.
+    # Splitting the catalog for memory's sake.
+    # Total memory footprint will be \approx Npus (on node) * (size of skymodel)
+    Nsrcs_total = len(catalog)
+    mem_avail = simutils.get_avail_memory()
+    Npus_node = mpi.get_node_comm().Get_size()
+    skymodel_mem_footprint = SkyModel._basesize * Nsrcs_total * Npus_node
+
+    if skymodel_mem_footprint > mem_avail:
+        Nsky_parts = np.ceil(skymodel_mem_footprint / float(mem_avail))
+        partsize = int(np.ceil(Nsrcs / Nsky_parts))
+        src_iter = [six.moves.range(s, s + partsize) for s in range(0, Nsrcs_total, partsize)]
+    else:
[…]
