### 2019-05-06 #1234 @ygrange
Hmkay, that's a bit frustrating. I will retry in the near future bur right now I don't have much time to do it and document it nicely and so on (sorry!), but what I did is build it in a clean ubuntu 18.04 docker container. The mpi I used is the default openmpi from that distro.  Could be the MPI build I guess.
