### 2019-05-07 1162e93
add pyuvdata to requirements.txt
