### 2019-05-07 a45c6fb
implement cleanup in tests
