### 2019-05-07 8e046b7
implement cleanup in tests
