+++ pyuvsim/tests/test_utils.py
@@ -16,6 +16,7 @@ import pyuvdata.tests as uvtest
 
 from pyuvsim.data import DATA_PATH as SIM_DATA_PATH
 from pyuvsim import utils as simutils
+import pyuvsim.tests as simtest
 
 triangle_uvfits_file = os.path.join(SIM_DATA_PATH, '28m_triangle_10time_10chan.uvfits')
 
@@ -99,10 +100,9 @@ def test_file_namer():
     """
     File name incrementer utility
     """
-    os.mkdir('tempfiles')
     fnames = []
     for i in range(111):
-        fname = os.path.join('tempfiles', 'file_' + str(i))
+        fname = os.path.join(simtestTESTDATA_PATH, 'file_' + str(i))
         with open(fname, 'w') as f:
             f.write(' ')
         fnames.append(fname)
@@ -110,7 +110,6 @@ def test_file_namer():
     new_filepath = simutils.check_file_exists_and_increment(existing_file)
     for fn in fnames:
         os.remove(fn)
-    os.rmdir('tempfiles')
     nt.assert_true(new_filepath.endswith("_111"))
 
 
@@ -118,16 +117,16 @@ def test_file_namer_extensions():
     """
     File name incrementer with specified extension
     """
-    os.mkdir('tempfiles')
+    os.mkdir(os.path.join(simtest.TESTDATA_PATH, 'tempfiles'))
     fnames = []
     for i in range(111):
-        fname = os.path.join('tempfiles', 'file_' + str(i) + '.ext')
+        fname = os.path.join(simtest.TESTDATA_PATH, 'tempfiles', 'file_' + str(i) + '.ext')
         with open(fname, 'w') as f:
             f.write(' ')
         fnames.append(fname)
     existing_file = fnames[0]
     new_filepath = simutils.check_file_exists_and_increment(existing_file, 'ext')
-    shutil.rmtree('tempfiles')
+    shutil.rmtree(os.path.join(simtest.TESTDATA_PATH, 'tempfiles'))
     nt.assert_true(new_filepath.endswith("_111.ext"))
 
 
@@ -138,7 +137,7 @@ def test_write_uvdata():
     uvtest.checkWarnings(uv.read_uvfits, [triangle_uvfits_file],
                          message='Telescope 28m_triangle_10time_10chan.yaml is not in known_telescopes.')
 
-    ofname = 'test_file'
+    ofname = os.path.join(simtest.TESTDATA_PATH, 'test_file')
     filing_dict = {'outfile_name': ofname}
     expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True)
     ofname = os.path.join('.', ofname)
@@ -146,9 +145,14 @@ def test_write_uvdata():
     expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True, out_format='miriad')
     nt.assert_equal(ofname, expected_ofname)
     nt.assert_raises(ValueError, simutils.write_uvdata, uv, filing_dict, return_filename=True, out_format='')
-    filing_dict['output_format'] = 'uvh5'
-    expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True)
-    nt.assert_equal(ofname + '.uvh5', expected_ofname)
+
+    try:
+        import h5py
+        filing_dict['output_format'] = 'uvh5'
+        expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True)
+        nt.assert_equal(ofname + '.uvh5', expected_ofname)
+    except ImportError:
+        pass # No h5py
 
     os.remove(ofname + '.uvh5')
     os.remove(ofname + '.uvfits')
