+++ pyuvsim/tests/__init__.py
@@ -13,6 +13,7 @@ from pyuvdata.data import DATA_PATH
 
 TESTDATA_PATH = 'temporary_test_data'
 
+
 def setup_package():
     """
     Make a directory for test data.
