### 2019-05-07 0fac1b6
link to comparison.rst from README
