### 2019-05-07 #1234 @aelanman
I tried an ubuntu 18.04 docker container with mpich and that worked, so it seems like it might be the open mpi build.

I had to remove pyuvdata from the requirements.txt because it broke the `pip install -r requirements.txt` build in the clean environment. I think the problem is that `pyuvdata` isn't really configured to be installed as a dependency via pip. I've been trying to work out what that issue is.

For now, though, I think I'll just add it to the installation instructions to install `pyuvdata` separately if installing with pip, and recommend the anaconda option once that's finished.
