### 2019-05-07 297f2a6
separate pyuvdata from requirements.txt and update installation instructions
