+++ requirements.txt
@@ -1,4 +1,3 @@
-pyuvdata>=1.3.7
 numpy>=1.15
 six>=1.11.0
 setuptools>=38.4.0
