### 2019-05-08 0f6ddd0
add new test of task index coverage
