### 2019-05-08 #1234 @aelanman
The requested changes have been merged to the master branch, so this is ready for a final review. @ygrange @arfon
