### 2019-05-09 de45c99
test source array splitting in low avail memory
