### 2019-05-09 #1234 @ygrange
The issue you report wrt pyuvdata doesn't sound completely unknown to me. For some reason I never mentioned it here but I think you are right: it somehow doesn't play well with pip.

My really absolutely final comment :)

Tests work almost like a charm using mpich, though I have one failure now in my setup:

```
yyy@017ab4273a6c:~/pyuvsim$ nosetests
............antenna_diameters is not set. Using known values for HERA.
..............................E................
ERROR: Test function that defines filenames from parameter dict
Traceback (most recent call last):
  File "/usr/lib/python2.7/dist-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/yyy/pyuvsim/pyuvsim/tests/test_utils.py", line 157, in test_write_uvdata
    os.remove(ofname + '.uvh5')
OSError: [Errno 2] No such file or directory: './temporary_test_data/test_file.uvh5'
-------------------- >> begin captured stdout << ---------------------
Outfile path:  ./temporary_test_data/test_file.uvfits
Outfile path:  ./temporary_test_data/test_file
Outfile path:  ./temporary_test_data/test_file_0

--------------------- >> end captured stdout << ----------------------

Ran 59 tests in 606.297s
```
Looking into the code it looks like line 157 cleans up no matter whether the try block in 149-153 did. I could probably even fix it myself and do a pull request (move line 157 inside the try block, or add a finally block), but I am not so sure that is ethically very clean.

Testing with h5py installed does work, as expected.
