### 2019-05-09 beb2175
minor bug in test_utils
