### 2019-05-09 #1234 @aelanman
Ah, sorry about that... I hadn't tried the tests without h5py. It should be redundant anyway. I've opened a PR with a fix. It should be merged within a couple hours.
