### 2019-05-09 #1234 @ygrange
Great!

Just for completeness: I advice the editor to proceed publishing!

--
Y. Grange

> Op 9 mei 2019 om 18:26 heeft Adam Lanman <notifications@github.com>
het volgende geschreven:
>
> Ah, sorry about that... I hadn't tried the tests without h5py. It
should be redundant anyway. I've opened a PR with a fix. It should be
merged within a couple hours.
>
