### 2019-05-09 b25e78d
restore test coverage
