### 2019-05-09 236697c
add checks of source cut calls in catalog readers
