+++ pyuvsim/tests/test_simsetup.py
@@ -14,7 +14,7 @@ import six
 import nose.tools as nt
 import astropy
 from astropy.time import Time
-from astropy.coordinates import Angle, SkyCoord, EarthLocation
+from astropy.coordinates import Angle, SkyCoord, EarthLocation, AltAz
 from astropy import units
 
 from pyuvdata import UVBeam, UVData
@@ -23,7 +23,6 @@ import pyuvdata.tests as uvtest
 import pyuvsim
 import pyuvsim.tests as simtest
 from pyuvsim.data import DATA_PATH as SIM_DATA_PATH
-import pyuvsim.tests as simtest
 
 herabeam_default = os.path.join(SIM_DATA_PATH, 'HERA_NicCST.uvbeam')
 
@@ -112,7 +111,7 @@ def test_catalog_from_params():
     nt.assert_true(np.all(catalog_str == catalog_uv))
 
 
-def test_flux_cuts():
+def test_param_flux_cuts():
     # Check that min/max flux limits in test params work.
 
     gleam_path = os.path.join(SIM_DATA_PATH, 'test_config', '..', 'gleam_50srcs.vot')
@@ -649,8 +648,81 @@ def test_point_catalog_reader():
         nt.assert_true(src.stokes[0] in catalog_table['flux_density_I'])
         nt.assert_true(src.freq.to("Hz").value in catalog_table['frequency'])
 
+    # Check cuts
+    source_select_kwds = {'min_flux': 1.0}
+    catalog = pyuvsim.simsetup.read_text_catalog(catfile, source_select_kwds=source_select_kwds, return_table=True)
+    nt.assert_true(len(catalog) == 2)
 
-def test_horizon_cut():
+
+def test_flux_cuts():
+    uv_in, beam_list, beam_dict = pyuvsim.simsetup.initialize_uvdata_from_params(manytimes_config)
+    Nsrcs = 20
+    uv_in.select(times=np.unique(uv_in.time_array)[:50], bls=[(0, 1)], metadata_only=True)
+
+    dt = np.format_parser(['U10', 'f8', 'f8', 'f8', 'f8'],
+                          ['source_id', 'ra_j2000', 'dec_j2000', 'flux_density_I', 'frequency'], [])
+
+    minflux = 0.5
+    maxflux = 3.0
+
+    catalog_table = np.recarray(Nsrcs, dtype=dt.dtype)
+    catalog_table['source_id'] = ["src{}".format(i) for i in range(Nsrcs)]
+    catalog_table['ra_j2000'] = np.random.uniform(0, 360., Nsrcs)
+    catalog_table['dec_j2000'] = np.linspace(-90, 90, Nsrcs)
+    catalog_table['flux_density_I'] = np.linspace(minflux, maxflux, Nsrcs)
+    catalog_table['frequency'] = np.ones(Nsrcs) * 200e6
+
+    minI_cut = 1.0
+    maxI_cut = 2.3
+
+    cut_sourcelist = pyuvsim.simsetup.source_cuts(catalog_table, min_flux=minI_cut, max_flux=maxI_cut)
+    nt.assert_true(np.all(cut_sourcelist['flux_density_I'] > minI_cut))
+    nt.assert_true(np.all(cut_sourcelist['flux_density_I'] < maxI_cut))
+
+
+def test_circumpolar_nonrising():
+    # Check that the source_cut function correctly identifies sources that are circumpolar or won't rise.
+    # Working with an observatory at the HERA latitude
+
+    lat = -31.0
+    lon = 0.0
+
+    Ntimes = 100
+    Nsrcs = 50
+
+    j2000 = 2451545.0
+    times = Time(np.linspace(j2000 - 0.5, j2000 + 0.5, Ntimes), format='jd', scale='utc')
+
+    ra = np.zeros(Nsrcs)
[…]
