+++ reference_simulations/obsparam_ref_1.3_gauss.yaml
@@ -1,11 +1,16 @@
 filing:
   outdir: "./simulation_results/"
   outfile_name: 'ref_1.3_gauss'
+  output_format: 'uvh5'
 freq:
-  Nfreqs:      6144
-  channel_width: 32552.0
-  start_freq:  50000000.0
-  end_freq:   250000000.0
+        #  Nfreqs:      6144
+        #  channel_width: 32552.0
+        #  start_freq:  50000000.0
+        #  end_freq:   250000000.0
+  Nfreqs:      6250.
+  channel_width: 32000.0
+  start_freq:  50016000.0
+  end_freq:   249984000.0
 sources:
   catalog: "catalog_files/letter_R_12pt_2458098.38824015.txt"
 telescope:
