+++ reference_simulations/obsparam_ref_1.1.yaml
@@ -1,6 +1,7 @@
 filing:
   outdir: "./simulation_results/"
   outfile_name: 'ref_1.1_uniform'
+  output_format: 'uvh5'
 freq:
   Nfreqs: 1
   channel_width: 80000.0
