### 2019-05-11 1099844
Update changelog and version number
