+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## [1.0.0] - 2019-5-10
 
 ### Added
