### 2019-05-13 b07c632
update warning catcher in tests
