### 2019-05-13 d0d37d5
update warning catcher in tests
