### 2019-05-14 392a658
use messages for all checkWarnings calls
