### 2019-05-14 fed5186
use messages for all checkWarnings calls
