### 2019-05-15 #1234 @aelanman
@arfon Okay. We've made new release and archived it on Zenodo here:
[https://doi.org/10.5281/zenodo.2847055](https://doi.org/10.5281/zenodo.2847055)
