### 2019-05-15 #1234 @arfon
@ygrange - many thanks for your review here ✨

@aelanman - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.01234 :zap::rocket::boom:
